% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/abi.R
\name{abi}
\alias{abi}
\alias{abi,FLStock,FLQuant-method}
\title{Calculate P obs for a FLStock object.}
\usage{
abi(object, age, ...)

\S4method{abi}{FLStock,FLQuant}(object, age)
}
\arguments{
\item{object}{A FLStock object.}

\item{age}{Reference ages obtained from abiAge.}

\item{...}{Additional arguments}
}
\value{
An FLQuant object containing P obs.
}
\description{
This function calculates P obs for a FLStock object.
}
