% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abi.R
\name{abiAge,FLBRP-method}
\alias{abiAge,FLBRP-method}
\title{Calculate Reference Age for FLBRP Object}
\usage{
\S4method{abiAge}{FLBRP}(object, ref = "msy", p = 0.9)
}
\arguments{
\item{object}{An FLBRP object.}

\item{ref}{Reference point, e.g., "msy" (default) or "f0.1".}

\item{p}{Probability threshold (default = 0.9).}
}
\value{
An FLQuant object containing reference ages.
}
\description{
Calculates the reference age for an FLBRP object, i.e., the age at which a given proportion (p) of the cumulative stock numbers is reached at a specified reference point (e.g., MSY).
}
