% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{abiAge}
\alias{abiAge}
\title{Calculate Process Error for FLBRP Object
Calculate the reference age for a FLBRP object.}
\usage{
abiAge(object, ref = "msy", p = 0.9)
}
\arguments{
\item{object}{FLBRP object.}

\item{ref}{Reference point, e.g., "msy" (default) or "f0.1".}

\item{p}{Probability threshold (default = 0.9).}
}
\value{
An FLQuant object containing reference ages.
}
\description{
This function calculates the reference age for a FLBRP object.
}
