% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abi.R
\name{abiMsy,FLBRP-method}
\alias{abiMsy,FLBRP-method}
\title{Calculate Proportion Above Reference Age at MSY for FLBRP Object}
\usage{
\S4method{abiMsy}{FLBRP}(object, ref = "msy", p = 0.9)
}
\arguments{
\item{object}{An FLBRP object.}

\item{ref}{Reference point, e.g., "msy" (default) or "f0.1".}

\item{p}{Probability threshold (default = 0.9).}
}
\value{
An FLQuant object containing the proportion above the reference age at MSY.
}
\description{
Calculates the proportion of stock numbers above the reference age at MSY for an FLBRP object.
}
