% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{abiMsy}
\alias{abiMsy}
\title{Calculate P(N) at FMSY for a FLBRP object.}
\usage{
abiMsy(object, ref = "msy", p = 0.9)
}
\arguments{
\item{ref}{Reference point, e.g., "msy" (default) or "f0.1".}

\item{p}{Probability threshold (default = 0.9).}

\item{y}{A FLBRP object.}
}
\value{
An FLQuant object containing P(N) at FMSY.
}
\description{
This function calculates P(N) at FMSY for a FLBRP object.
}
