% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abi.R
\name{abistock}
\alias{abistock}
\title{Helper: Proportion Above Reference Age for FLStock}
\usage{
abistock(x, A)
}
\arguments{
\item{x}{An FLStock object.}

\item{A}{Reference age (FLQuant).}
}
\value{
An FLQuant object with the proportion above the reference age.
}
\description{
Calculates the proportion of stock numbers above a given reference age for an FLStock object.
}
\keyword{internal}
