% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario-management.R
\name{applyPriors}
\alias{applyPriors}
\title{Apply Priors to FLBRP Object}
\usage{
applyPriors(flbrp, prior_config)
}
\arguments{
\item{flbrp}{FLBRP object}

\item{prior_config}{Prior configuration}
}
\value{
FLBRP object with applied priors
}
\description{
Apply prior configuration to FLBRP object.
}
