% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario-management.R
\name{applyScenarioConfig}
\alias{applyScenarioConfig}
\title{Apply Scenario Configuration}
\usage{
applyScenarioConfig(flbrp_list, scenario_config, analysis_function)
}
\arguments{
\item{flbrp_list}{List of FLBRP objects}

\item{scenario_config}{Scenario configuration object}

\item{analysis_function}{Function to apply with scenario configuration}
}
\value{
Results of applying scenario configuration
}
\description{
Apply scenario configuration to FLBRP objects.
}
