% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior-thread.R
\name{asJabbaPriors}
\alias{asJabbaPriors}
\title{Convert Canonical Priors to JABBA Prior Object}
\usage{
asJabbaPriors(canonical)
}
\arguments{
\item{canonical}{Canonical prior data.frame.}
}
\value{
A JABBA prior object with `pr` and `pr.sd` as `FLPar`.
}
\description{
Convert Canonical Priors to JABBA Prior Object
}
