% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-generic.R, R/ices-attributes.R
\name{benchmark}
\alias{benchmark}
\alias{benchmark,FLStock-method}
\alias{benchmark,FLStocks-method}
\alias{benchmark,list-method}
\alias{benchmark,FLBRP-method}
\title{Extract ICES benchmark reference points from an FLStock}
\usage{
benchmark(object, ...)

\S4method{benchmark}{FLStock}(object)

\S4method{benchmark}{FLStocks}(object)

\S4method{benchmark}{list}(object)

\S4method{benchmark}{FLBRP}(object)
}
\arguments{
\item{object}{An \code{FLStock}, \code{FLStocks}, \code{FLBRP}, or list}

\item{...}{Unused}
}
\value{
An \code{FLPar} (single stock) or a data.frame (several stocks)
}
\description{
Extract ICES benchmark reference points from an FLStock
}
