% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr-rs.R
\name{bevertonRec}
\alias{bevertonRec}
\alias{bevertonRec,FLPar,numeric-method}
\alias{bevertonRec,FLPar,FLPar-method}
\alias{bevertonRec,FLPar,FLQuant-method}
\title{Generic for SSB at a given recruitment level for Beverton-Holt model}
\usage{
bevertonRec(params, rec, ...)

\S4method{bevertonRec}{FLPar,numeric}(params, rec, ...)

\S4method{bevertonRec}{FLPar,FLPar}(params, rec, ...)

\S4method{bevertonRec}{FLPar,FLQuant}(params, rec, ...)
}
\description{
Generic for SSB at a given recruitment level for Beverton-Holt model
}
