% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr-rs.R
\name{bevholtInv}
\alias{bevholtInv}
\title{Inverse Beverton-Holt stock-recruitment relationship}
\usage{
bevholtInv(params, rec)
}
\arguments{
\item{params}{FLPar with parameters 'a' and 'b'}

\item{rec}{Recruitment values}
}
\value{
SSB values
}
\description{
Inverse Beverton-Holt stock-recruitment relationship
}
\keyword{internal}
