% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{blim}
\alias{blim}
\title{Calculate Blim Reference Point}
\usage{
blim(object, ...)
}
\arguments{
\item{object}{An FLBRP object}

\item{ratio}{Ratio of virgin recruitment to use (default 0.3)}
}
\value{
An FLPar object containing Blim reference points
}
\description{
Calculate Blim Reference Point
}
