% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pella_tomlinson.R
\name{bmsy,FLPar-method}
\alias{bmsy,FLPar-method}
\title{Method for BMSY (FLPar)}
\usage{
\S4method{bmsy}{FLPar}(params)
}
\arguments{
\item{params}{FLPar object.}
}
\value{
Numeric value or FLQuant object.
}
\description{
Method for calculating BMSY when parameters are provided as an FLPar object.
}
