% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pella_tomlinson.R
\name{bmsy,numeric-method}
\alias{bmsy,numeric-method}
\title{Method for BMSY (numeric)}
\usage{
\S4method{bmsy}{numeric}(params)
}
\arguments{
\item{params}{Numeric vector.}
}
\value{
Numeric value.
}
\description{
Method for calculating BMSY when parameters are provided as a numeric vector.
}
