% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pella_tomlinson.R
\name{bmsy,PellaTomlinson-method}
\alias{bmsy,PellaTomlinson-method}
\title{Method for BMSY (PellaTomlinson)}
\usage{
\S4method{bmsy}{PellaTomlinson}(params)
}
\arguments{
\item{params}{PellaTomlinson object.}
}
\value{
Numeric value or FLQuant object.
}
\description{
Method for calculating BMSY when the PellaTomlinson object is used.
}
