% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pella_tomlinson.R
\name{bmsy}
\alias{bmsy}
\title{Calculate Biomass at MSY (BMSY)}
\usage{
bmsy(params, ...)

bmsy(params, ...)
}
\arguments{
\item{params}{FLPar object.}

\item{...}{Additional arguments.}
}
\value{
Numeric value of BMSY.

Numeric value or FLQuant object.
}
\description{
Calculates the Biomass at MSY (BMSY) given model parameters.

Generic function for calculating Biomass at MSY (BMSY).
}
\examples{
params = c(p = 0.25, virgin = 1.0)
bmsy_val = bmsy(params)

}
