% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pellatUtils.R
\name{bmsyK}
\alias{bmsyK}
\title{Calculate BMSY/K ratio from shape parameter p}
\usage{
bmsyK(p)
}
\arguments{
\item{p}{Shape parameter of Pella-Tomlinson model}
}
\value{
BMSY/K ratio
}
\description{
Calculate BMSY/K ratio from shape parameter p
}
