% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pellatUtils.R
\name{bmsyK2p}
\alias{bmsyK2p}
\title{Robust root-finding function for p given a valid BMSY/K ratio}
\usage{
bmsyK2p(target_ratio, lower = 0.05, upper = 5, tol = 1e-08)
}
\arguments{
\item{target_ratio}{BMSY/K ratio (must be between 0 and 1)}

\item{lower}{Lower bound for p search}

\item{upper}{Upper bound for p search}

\item{tol}{Tolerance for root finding}
}
\value{
Shape parameter p
}
\description{
Robust root-finding function for p given a valid BMSY/K ratio
}
