% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrapFLBRP}
\alias{bootstrapFLBRP}
\title{Bootstrap Analysis Functions for FLRebuild Package}
\usage{
bootstrapFLBRP(
  flbrp_list,
  nits = 100,
  ftmb_args = list(),
  prior_data = NULL,
  ncores = NULL
)
}
\arguments{
\item{flbrp_list}{List of FLBRP objects to bootstrap}

\item{nits}{Number of bootstrap iterations (default=100)}

\item{ftmb_args}{List of arguments to pass to ftmb3 function}

\item{prior_data}{Data frame with stock-specific priors (optional)}

\item{ncores}{Number of cores to use (default=detectCores()-4)}
}
\value{
List of bootstrapped FLBRP objects
}
\description{
This file contains functions for performing bootstrap analysis on FLBRP objects
with parallel processing capabilities.
Bootstrap FLBRP Objects
}
\details{
Perform bootstrap analysis on FLBRP objects with parallel processing.
}
