% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrapSingleFLBRP}
\alias{bootstrapSingleFLBRP}
\title{Bootstrap Single FLBRP Object}
\usage{
bootstrapSingleFLBRP(flbrp, nits = 100, ftmb_args = list(), prior_data = NULL)
}
\arguments{
\item{flbrp}{Single FLBRP object to bootstrap}

\item{nits}{Number of bootstrap iterations}

\item{ftmb_args}{List of arguments to pass to ftmb3 function}

\item{prior_data}{Data frame with stock-specific priors (optional)}
}
\value{
Bootstrapped FLBRP object
}
\description{
Perform bootstrap analysis for a single FLBRP object.
}
