% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pellatRebuild.R
\name{brebuild}
\alias{brebuild}
\alias{brebuild,numeric-method}
\alias{brebuild,FLPar-method}
\alias{brebuild,FLBRP-method}
\title{Calculate initial biomass required for recovery to BMSY in given time}
\usage{
brebuild(object, ...)

\S4method{brebuild}{numeric}(object, r, p, bmsy = 1, F = 0, ...)

\S4method{brebuild}{FLPar}(object, F = 0, ...)

\S4method{brebuild}{FLBRP}(object, nyrs, F = 0, biomass = "ssb", ...)
}
\arguments{
\item{...}{Additional arguments}

\item{r}{Intrinsic rate of population increase}

\item{p}{Shape parameter of the Pella-Tomlinson model}

\item{bmsy}{Biomass at maximum sustainable yield (default=1)}

\item{F}{Fishing mortality rate during recovery (default=0)}

\item{nyrs}{Number of years for recovery}
}
\value{
Initial biomass required for recovery
}
\description{
Numerically integrates the Pella-Tomlinson production function to estimate the initial biomass 
required for a population to recover to BMSY within a specified number of years, with optional fishing mortality during recovery.
}
\examples{
brebuild(10, r=0.2, p=1.5)
brebuild(10, r=0.2, p=1.5, F=0.05)

# With FLPar
library(FLCore)
params=FLPar(nyrs=10, r=0.2, p=1.5, bmsy=1)
brebuild(params, F=0.05)

# With FLBRP object
data(ple4brp)
brebuild(ple4brp, nyrs=10, F=0.05)
}
