% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLBRP-priors.R
\name{calcPriors}
\alias{calcPriors}
\alias{calcPriors,FLBRP-method}
\alias{calcPriors,FLStock-method}
\alias{calcPriors,FLStocks-method}
\title{Calculate priors for biomass dynamic models}
\usage{
calcPriors(object, ...)

\S4method{calcPriors}{FLBRP}(object)

\S4method{calcPriors}{FLStock}(object, nmin = 0:2, nmax = 0:2)

\S4method{calcPriors}{FLStocks}(object, nmin = 0:2, nmax = 0:2)
}
\arguments{
\item{object}{An object of class FLBRP, FLStock, or FLStocks}

\item{nmin}{Numeric vector for minimum years (for FLStock method)}

\item{nmax}{Numeric vector for maximum years (for FLStock method)}
}
\value{
An FLPar object (for FLBRP) or data.frame (for FLStock/FLStocks) containing prior parameters
}
\description{
Calculates priors for Pella-Tomlinson production models from FLBRP, FLStock, or FLStocks objects
}
