% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculateRecoveryTime}
\alias{calculateRecoveryTime}
\title{Calculate recovery time based on population growth rate}
\usage{
calculateRecoveryTime(initial_biomass, target_biomass, growth_rate)
}
\arguments{
\item{initial_biomass}{Initial biomass relative to target}

\item{target_biomass}{Target biomass (e.g., BMSY)}

\item{growth_rate}{Population growth rate}
}
\value{
Time to recovery in years
}
\description{
Calculate time to recovery using population growth rate
}
