% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-utils.R
\name{calculateStatistics}
\alias{calculateStatistics}
\title{Calculate Statistics}
\usage{
calculateStatistics(x, statistics = c("mean", "median", "sd", "min", "max"))
}
\arguments{
\item{x}{Numeric vector}

\item{statistics}{Character vector of statistics to calculate}
}
\value{
Named vector of statistics
}
\description{
Calculate specified statistics for a numeric vector.
}
