% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculateTmax}
\alias{calculateTmax}
\title{Calculate Tmax using different methods}
\usage{
calculateTmax(object, method = "generation", tmin = NULL, mfmt = NULL)
}
\arguments{
\item{object}{An FLR object with life history parameters}

\item{method}{Method to use: "generation" (Tmin + generation time), 
"fishing" (time to rebuild at 75% MFMT), or "multiply" (Tmin * 2)}

\item{tmin}{Minimum rebuilding time}

\item{mfmt}{Maximum fishing mortality threshold}
}
\value{
Tmax value
}
\description{
Calculate maximum rebuilding time (Tmax) using various methods
as described in NS1 guidelines
}
