% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refpt-processing.R
\name{cleanColumnNames}
\alias{cleanColumnNames}
\title{Clean Column Names}
\usage{
cleanColumnNames(data)
}
\arguments{
\item{data}{Data frame with messy column names}
}
\value{
Data frame with clean column names
}
\description{
Clean up column names in merged data.
}
