% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R, R/parallel-utils.R
\name{cleanupParallelProcessing}
\alias{cleanupParallelProcessing}
\title{Cleanup Parallel Processing}
\usage{
cleanupParallelProcessing(cl)

cleanupParallelProcessing(cl)
}
\arguments{
\item{cl}{Parallel processing cluster}
}
\description{
Clean up parallel processing resources.

Clean up parallel processing resources.
}
