% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refpt-processing.R
\name{combineScenarioResults}
\alias{combineScenarioResults}
\title{Combine Scenario Results}
\usage{
combineScenarioResults(scenario_results, scenario_names, srr_names)
}
\arguments{
\item{scenario_results}{List of scenario results}

\item{scenario_names}{Character vector of scenario names}

\item{srr_names}{Character vector of SRR names}
}
\value{
Combined results data frame
}
\description{
Combine results from multiple scenarios into a single data frame.
}
