% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario-management.R
\name{compareMultipleScenarios}
\alias{compareMultipleScenarios}
\title{Compare Multiple Scenarios}
\usage{
compareMultipleScenarios(scenario_results)
}
\arguments{
\item{scenario_results}{List of scenario results}
}
\value{
Comparison results
}
\description{
Compare multiple scenarios using statistical tests.
}
