% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-plots.R
\name{createSummaryStatsPlot}
\alias{createSummaryStatsPlot}
\title{Create Summary Statistics Plot}
\usage{
createSummaryStatsPlot(
  data,
  group_var,
  value_var,
  stat_type = "mean",
  plot_type = "bar"
)
}
\arguments{
\item{data}{Data for summary statistics}

\item{group_var}{Variable to group by}

\item{value_var}{Variable to summarize}

\item{stat_type}{Type of statistic ("mean", "median", "sd", "range")}

\item{plot_type}{Type of plot ("bar", "line", "point")}
}
\value{
ggplot object
}
\description{
Create summary statistics visualization.
}
