% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mFn.R
\name{ddM}
\alias{ddM}
\alias{ddM,FLQuant,FLPar-method}
\title{Calculate Density-Dependent Mortality}
\usage{
ddM(wt, par)

\S4method{ddM}{FLQuant,FLPar}(wt, par)
}
\arguments{
\item{wt}{An FLQuant object containing weights}

\item{par}{An FLPar object containing parameters 'm1' and 'm2'}
}
\value{
An FLQuant object with density-dependent mortality rates
}
\description{
Calculates density-dependent mortality rate based on weight and parameters.
Uses the formula: M = m1 * (wt^m2), where m1 and m2 are parameters.
}
\details{
This function calculates mortality as a power function of weight:
M = m1 * (wt^m2)
where m1 is the scaling parameter and m2 is the exponent.
}
\examples{
\dontrun{
library(FLCore)
wt=FLQuant(c(0.1, 0.5, 1.0, 2.0))
par=FLPar(m1 = 0.3, m2 = -0.3)
ddM(wt, par)
}

}
