% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr-rs.R
\name{.bevertonRec_numeric}
\alias{.bevertonRec_numeric}
\title{Helper function for bevertonRec with numeric inputs}
\usage{
.bevertonRec_numeric(params, rec)
}
\description{
Helper function for bevertonRec with numeric inputs
}
\keyword{internal}
