% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr-rs.R
\name{.rickerRec_numeric}
\alias{.rickerRec_numeric}
\title{Helper function for rickerRec with numeric inputs}
\usage{
.rickerRec_numeric(params, rec)
}
\description{
Helper function for rickerRec with numeric inputs
}
\keyword{internal}
