% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr-rs.R
\name{.segRegRec_numeric}
\alias{.segRegRec_numeric}
\title{Helper function for segRegRec with numeric inputs}
\usage{
.segRegRec_numeric(params, rec)
}
\description{
Helper function for segRegRec with numeric inputs
}
\keyword{internal}
