% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-generic.R, R/ices-attributes.R
\name{eqsim}
\alias{eqsim}
\alias{eqsim,FLStock-method}
\alias{eqsim,FLStocks-method}
\alias{eqsim,list-method}
\title{Extract EqSim reference points from an FLStock}
\usage{
eqsim(object, ...)

\S4method{eqsim}{FLStock}(object)

\S4method{eqsim}{FLStocks}(object)

\S4method{eqsim}{list}(object)
}
\arguments{
\item{object}{An \code{FLStock}, \code{FLStocks}, or list of stocks}

\item{...}{Unused}
}
\value{
An \code{FLPar} (single stock) or a data.frame (several stocks)
}
\description{
Extract EqSim reference points from an FLStock
}
