% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pella_tomlinson.R
\name{estimatePellaTomlinsonParams}
\alias{estimatePellaTomlinsonParams}
\title{Utility function for estimating Pella-Tomlinson parameters from bmsy, virgin, and msy}
\usage{
estimatePellaTomlinsonParams(bmsy, virgin, msy)
}
\arguments{
\item{bmsy}{Numeric vector of BMSY values.}

\item{virgin}{Numeric vector of virgin biomass values.}

\item{msy}{Numeric vector of MSY values.}
}
\value{
Data frame with estimated r and p values.
}
\description{
Estimates Pella-Tomlinson parameters (r, p) from observed BMSY, virgin, and MSY.
This function is designed to work with ddply and other data manipulation functions.
}
\examples{
bmsy_vals = c(0.7, 0.5, 0.8, 0.3)
virgin_vals = c(1.0, 1.0, 1.0, 1.0)
msy_vals = c(0.1, 0.08, 0.12, 0.06)

estimated_params = estimatePellaTomlinsonParams(bmsy_vals, virgin_vals, msy_vals)

}
