% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebuild-analysis.R
\name{evaluateRebuildPotential}
\alias{evaluateRebuildPotential}
\title{Evaluate Rebuild Potential Across Scenarios}
\usage{
evaluateRebuildPotential(rebuild_time_data, gen_time_data, scenarios, srrs)
}
\arguments{
\item{rebuild_time_data}{Rebuild time data}

\item{gen_time_data}{Generation time data}

\item{scenarios}{Character vector of scenario names}

\item{srrs}{Character vector of SRR names}
}
\value{
Rebuild potential assessment
}
\description{
Evaluate rebuild potential across different scenarios.
}
