% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-utils.R
\name{exportToPDF}
\alias{exportToPDF}
\title{Export to PDF Format}
\usage{
exportToPDF(results, filename, title = "Analysis Results")
}
\arguments{
\item{results}{Results object}

\item{filename}{Output filename}

\item{title}{PDF title}
}
\value{
PDF file path
}
\description{
Export results to PDF format using rmarkdown.
}
