% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refpt-processing.R
\name{extractRefpts}
\alias{extractRefpts}
\title{Reference Point Processing Functions for FLRebuild Package}
\usage{
extractRefpts(flbrp_list, refpts_to_extract = c("msy", "f0.1", "spr.30"))
}
\arguments{
\item{flbrp_list}{List of FLBRP objects}

\item{refpts_to_extract}{Character vector of reference points to extract (default: c("msy", "f0.1", "spr.30"))}
}
\value{
Data frame of extracted reference points
}
\description{
This file contains functions for extracting and processing reference points
from FLBRP objects.
Extract Reference Points from FLBRP Objects
}
\details{
Extract reference points from a list of FLBRP objects.
}
