% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStock-indicatorsBiol.R
\name{fapex,FLBRP-method}
\alias{fapex,FLBRP-method}
\title{Calculate Apical Fishing Mortality}
\usage{
\S4method{fapex}{FLBRP}(x, ...)
}
\arguments{
\item{x}{An FLBRP object}

\item{...}{Additional arguments}
}
\value{
An FLQuant object with apical fishing mortality values
}
\description{
Calculates the maximum fishing mortality across ages (apical F) for an FLBRP object.
}
