% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-generic.R, R/ices-attributes.R
\name{fishlife}
\alias{fishlife}
\alias{fishlife,FLStock-method}
\alias{fishlife,FLStocks-method}
\alias{fishlife,list-method}
\title{Extract FishLife parameters from an FLStock}
\usage{
fishlife(object, ...)

\S4method{fishlife}{FLStock}(object)

\S4method{fishlife}{FLStocks}(object)

\S4method{fishlife}{list}(object)
}
\arguments{
\item{object}{An \code{FLStock}, \code{FLStocks}, or list of stocks}

\item{...}{Unused}
}
\value{
An \code{FLPar} (single stock) or a data.frame (several stocks)
}
\description{
Extract FishLife parameters from an FLStock
}
