% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flicc-methods.R
\name{fishLifeParsByStock}
\alias{fishLifeParsByStock}
\alias{fishlifeParsByStock}
\title{FishLife Parameter Interface by Stock}
\usage{
fishLifeParsByStock(
  spp,
  stockCol = "stock",
  genusCol = "Genus",
  speciesCol = "Species"
)

fishlifeParsByStock(
  spp,
  stockCol = "stock",
  genusCol = "Genus",
  speciesCol = "Species"
)
}
\arguments{
\item{spp}{A data.frame with at least stock, genus, species columns.}

\item{stockCol}{Name of stock column.}

\item{genusCol}{Name of genus column.}

\item{speciesCol}{Name of species column.}
}
\value{
A named list of \code{FLPar} (`linf`, `k`, `t0`, `l50`, `m`) by stock.
  If no FishLife match is found for a stock, values are returned as `NA`.
}
\description{
Replicates the prototype `ParFishlife <- lapply(split(spp, spp$stock), ...)`
workflow in a reusable function.
}
