% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flicc-methods.R
\name{fitFLICC}
\alias{fitFLICC}
\alias{fitFLICC,FLStock-method}
\alias{fitFLICC,FLStocks-method}
\title{Fit FLICC Models from FLR Stocks}
\usage{
fitFLICC(object, ...)

\S4method{fitFLICC}{FLStock}(
  object,
  lengthData = NULL,
  nLengths = 100,
  catch = 12000,
  mk = c(2, 0.01),
  selFun = "logistic",
  refLength = 5,
  compile = TRUE,
  modelName = "Base: Mk ~ Length-inverse",
  ...
)

\S4method{fitFLICC}{FLStocks}(object, ...)
}
\arguments{
\item{object}{Stock object.}

\item{...}{Additional method arguments.}

\item{lengthData}{Optional length-frequency FLQuant. If NULL, the method
tries to build one via \code{icesdata::lenSamples()}.}

\item{nLengths}{Number of simulated fish for \code{lenSamples()}.}

\item{catch}{Total catch value used in FLICC input.}

\item{mk}{Initial \code{Mk} parameters passed to FLICC.}

\item{selFun}{Selectivity function name.}

\item{refLength}{Reference length passed to FLICC.}

\item{compile}{Logical passed to \code{FLicc::fiticc()}.}

\item{modelName}{Label for the FLICC model.}
}
\value{
A list with estimated parameters, model input and fit output.
}
\description{
Generic interface to prepare FLICC inputs from FLR stock objects and fit
\code{FLicc::fiticc()}.
}
