% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pellat.R
\name{fitPellatDirect}
\alias{fitPellatDirect}
\title{Direct calculation of Pella-Tomlinson parameters}
\usage{
fitPellatDirect(msy, bmsy, virgin, ...)
}
\arguments{
\item{msy}{Maximum sustainable yield}

\item{bmsy}{Biomass at MSY}

\item{virgin}{Virgin biomass}

\item{...}{Additional arguments}
}
\value{
List with estimated parameters
}
\description{
Direct calculation of Pella-Tomlinson parameters
}
\examples{
result = fitPellatDirect(msy=100, bmsy=500, virgin=1000)
}
