% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pella_tomlinson.R
\name{FLPar-methods}
\alias{FLPar-methods}
\title{Methods for FLPar objects}
\description{
Methods for calculating production, MSY, BMSY, FMSY, p, r, param
when parameters are provided as FLPar objects.
}
\keyword{internal}
