% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pella_tomlinson.R
\name{fmsy,FLPar-method}
\alias{fmsy,FLPar-method}
\title{Method for FMSY (FLPar)}
\usage{
\S4method{fmsy}{FLPar}(params)
}
\arguments{
\item{params}{FLPar object.}
}
\value{
Numeric value or FLQuant object.
}
\description{
Method for calculating FMSY when parameters are provided as an FLPar object.
}
