% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pella_tomlinson.R
\name{fmsy,numeric-method}
\alias{fmsy,numeric-method}
\title{Method for FMSY (numeric)}
\usage{
\S4method{fmsy}{numeric}(params)
}
\arguments{
\item{params}{Numeric vector.}
}
\value{
Numeric value.
}
\description{
Method for calculating FMSY when parameters are provided as a numeric vector.
}
