% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pella_tomlinson.R
\name{fmsy,PellaTomlinson-method}
\alias{fmsy,PellaTomlinson-method}
\title{Method for FMSY (PellaTomlinson)}
\usage{
\S4method{fmsy}{PellaTomlinson}(params)
}
\arguments{
\item{params}{PellaTomlinson object.}
}
\value{
Numeric value or FLQuant object.
}
\description{
Method for calculating FMSY when the PellaTomlinson object is used.
}
