% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pella_tomlinson.R
\name{fmsy}
\alias{fmsy}
\title{Calculate Fishing Mortality at MSY (FMSY)}
\usage{
fmsy(params, ...)

fmsy(params, ...)
}
\arguments{
\item{params}{FLPar object.}

\item{...}{Additional arguments.}
}
\value{
Numeric value of FMSY.

Numeric value or FLQuant object.
}
\description{
Calculates the Fishing Mortality at MSY (FMSY) given model parameters.

Generic function for calculating Fishing Mortality at MSY (FMSY).
}
\examples{
params = c(r = 0.5, p = 0.25)
fmsy_val = fmsy(params)

}
