% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mFn.R
\name{forage}
\alias{forage}
\alias{forage,FLQuant-method}
\alias{forage,FLStock-method}
\alias{forage,FLBRP-method}
\title{Calculate Forage Biomass}
\usage{
forage(x)

\S4method{forage}{FLQuant}(x)

\S4method{forage}{FLStock}(x)

\S4method{forage}{FLBRP}(x)
}
\arguments{
\item{x}{An FLQuant object or an object with stock data (e.g., FLStock, FLBRP)}
}
\value{
An FLQuant object with forage biomass by year and iteration
}
\description{
Calculates forage biomass, representing the biomass available to predators,
weighted by mortality components and total mortality.
}
\details{
Forage biomass is calculated as:
Forage = sum over ages of [stock.wt * stock.n * (M - M1) / Z * (1 - exp(-Z))]
where:
- M is natural mortality
- M1 is background mortality
- Z is total mortality (M + F)

This represents the biomass available to predators, accounting for
the variable mortality component (M2 = M - M1) and mortality rates.

For FLQuant objects, this method is not directly applicable as it requires
multiple stock components (stock.wt, stock.n, m, z). Use with FLStock or FLBRP objects instead.
}
\examples{
\dontrun{
library(FLCore)
library(FLBRP)
data(ple4brp)
forage_bio=forage(ple4brp)
}

}
