% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fromLogits}
\alias{fromLogits}
\title{Convert Logit-Transformed Steepness to Original Scale}
\usage{
fromLogits(logitH)
}
\arguments{
\item{logitH}{Numeric vector of logit-transformed steepness values}
}
\value{
Numeric vector of steepness values on original scale (0.2 to 1.0)
}
\description{
Converts steepness parameter from logit-transformed scale back to the original
scale used in fisheries models.
}
\examples{
# Convert logit steepness values to original scale (0.2 to 1.0)
logit_values <- c(0, 1, 2)
steepness <- fromLogits(logit_values)
print(steepness)

}
