% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftmb.R
\name{ftmb2}
\alias{ftmb2}
\title{Fits Stock Recruitment Relationships (SRR) in TBM (with flexible priors for s and r0)}
\usage{
ftmb2(
  object,
  spr0 = spr0,
  model = SRModelName(FLCore::model(object)),
  s = 0.7,
  s.est = TRUE,
  s.logitsd = 1.3,
  inflect = NA,
  inits = function(object = object, s = s, inflect = inflect) {
     inflect_init <-
    ifelse(is.na(inflect), 0, log(inflect))
     c(median(c(log(rec(object)))), log(0.4),
    toLogits(s), inflect_init)
 },
  lower = function(object = object, s = -20, inflect = inflect) {
     inflect_lower <-
    ifelse(is.na(inflect), -20, log(inflect))
     c(quantile(c(log(rec(object))), probs
    = 0.1, na.rm = TRUE), log(0.05), -20, inflect_lower)
 },
  upper = function(object = object, s = 20, inflect = inflect) {
     inflect_upper <-
    ifelse(is.na(inflect), 20, log(inflect))
     c(quantile(c(log(rec(object))), probs =
    0.9, na.rm = TRUE), log(1.5), 20, inflect_upper)
 },
  short = TRUE,
  SDreport = short,
  prior_s = NULL,
  cv_s = NULL,
  prior_r0 = NULL,
  cv_r0 = NULL
)
}
\arguments{
\item{object}{Input FLSR object.}

\item{spr0}{unfished spawning biomass per recruit from spr0Yr(FLStock)}

\item{s}{steepness parameter of SRR (fixed or prior mean)}

\item{s.est}{option to estimate steepness}

\item{s.logitsd}{prior sd for logit(s), default is 1.3 (flat) if s.est = TRUE}

\item{inflect}{Inflection point for the segreg model. If NA (default), it is estimated; if numeric, it is fixed at the provided value.}

\item{inits}{option to specify initial values of log(r0), log(SigR) and logit(s)}

\item{lower}{option to specify lower bounds of log(r0), log(SigR) and logit(s)}

\item{upper}{option to specify upper bounds of log(r0), log(SigR) and logit(s)}

\item{SDreport}{option to converge hessian and get vcov}

\item{prior_s}{Optional prior mean for steepness (s)}

\item{cv_s}{Optional coefficient of variation for steepness prior}

\item{prior_r0}{Optional prior mean for r0}

\item{cv_r0}{Optional coefficient of variation for r0 prior}
}
\value{
A list containing elements 'FLSR', of class *FLSR*
}
\description{
Not exported. Use \code{icesdata::ftmb2} for standard fits. This copy
exists so \code{ftmb_b0dyn} can use the FLRebuild TMB DLL (\code{use_b0}).
}
\examples{
# See ftmb for usage

}
