% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebuild_biodyn.R
\name{fwdPT}
\alias{fwdPT}
\title{Forward Projection Function for Rebuild Time Calculation}
\usage{
fwdPT(object, nyrs = 50, ftar = 0, timing = 0)
}
\arguments{
\item{object}{FLPar object with parameters}

\item{nyrs}{Number of years for projection}

\item{ftar}{Target fishing mortality multiplier}

\item{timing}{Timing parameter}

\item{...}{Additional arguments}
}
\value{
Data frame with rebuild time results
}
\description{
Internal function that performs forward projection using Pella-Tomlinson dynamics.
This is the core calculation engine for rebuildTime3.
}
\keyword{internal}
