% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-utils.R
\name{generateReportContent}
\alias{generateReportContent}
\title{Generate Report Content}
\usage{
generateReportContent(analysis_results, report_config)
}
\arguments{
\item{analysis_results}{Analysis results}

\item{report_config}{Report configuration}
}
\value{
Report content
}
\description{
Generate content for analysis report.
}
