% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pellatParams.R
\name{getComparisonTable}
\alias{getComparisonTable}
\title{Get comparison matrix}
\usage{
getComparisonTable(object, digits = 6)
}
\arguments{
\item{object}{A pellatParams object}

\item{digits}{Number of digits to display (default: 6)}
}
\value{
A formatted matrix
}
\description{
Returns a clean comparison matrix between observed and estimated parameters.
}
\examples{
pt = pellatParams(c(msy = 500, fmsy = 0.3, virgin = 10000))
pt = estimateParams(pt)
getComparisonTable(pt)
}
