% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMstats.R
\name{getPriors}
\alias{getPriors}
\title{Get priors from FLBRP object}
\usage{
getPriors(x)
}
\arguments{
\item{x}{An FLBRP object}
}
\value{
A data.frame containing prior parameters
}
\description{
Extracts prior information from an FLBRP object including reference points and biomass/fishing mortality values
}
