% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvln.R
\name{getRuns}
\alias{getRuns}
\alias{getRuns,character-method}
\alias{getRuns,list-method}
\title{Load Stock Synthesis projection runs}
\usage{
getRuns(object, ...)

\S4method{getRuns}{character}(object, pattern = "_([0-9]+)t\\\\.Rdata$", ...)

\S4method{getRuns}{list}(object, ...)
}
\arguments{
\item{object}{Named character vector of run folders or \code{.rds}/\code{.Rdata}
paths, a named list of \code{SS_output} objects, or a directory whose
\code{*.Rdata} files are named \code{*_\\{catch\\}t.Rdata}.}

\item{...}{Unused.}
}
\value{
A named list of \code{SS_output}-shaped objects; names are catch
  levels coercible to numeric (required by \code{\link{kobeMomentGrid}}).
}
\description{
Load Stock Synthesis projection runs
}
\examples{
\dontrun{
runs <- getRuns(c(
  "0" = "path/to/HW2e_..._0t.Rdata",
  "250" = "path/to/HW2e_..._250t.Rdata"))
runs <- getRuns("P:/.../SSoutput")  # discovers *_\\{catch\\}t.Rdata
}
}
\seealso{
\code{\link{kobeMomentGrid}}, \code{\link{makeK2SM}}
}
