% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodyn-rebuild.R
\name{gt}
\alias{gt}
\alias{gt,Rebuild-method}
\title{Generation Time}
\usage{
gt(x)

\S4method{gt}{Rebuild}(x)
}
\arguments{
\item{x}{An object of class \code{Rebuild}}
}
\value{
Numeric value representing generation time
}
\description{
Calculates generation time as 1/r
}
