% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pellatRebuild.R
\name{gtLife}
\alias{gtLife}
\title{Calculate generation time from a life table}
\usage{
gtLife(ages, lx, mx)
}
\arguments{
\item{ages}{Vector of ages}

\item{lx}{Vector of age-specific survivorship (proportion surviving to each age)}

\item{mx}{Vector of age-specific fecundity (mean number of offspring at each age)}
}
\value{
Generation time in years
}
\description{
Calculates the mean generation time from age-specific survivorship and fecundity.
}
\examples{
ages<-0:5
lx<-c(1, 0.8, 0.6, 0.4, 0.2, 0.1)
mx<-c(0, 0, 2, 3, 1, 0)
gtLife(ages, lx, mx)
}
