% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebuild-methods.R
\name{interp}
\alias{interp}
\title{Interpolate rebuilding times}
\usage{
interp(df)
}
\arguments{
\item{df}{Data frame with columns: initial, year, ssb}
}
\value{
Data frame with columns: initial, year
}
\description{
Interpolate rebuilding times
}
\keyword{internal}
