% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pellatParams.R
\name{isValidPellatParams}
\alias{isValidPellatParams}
\title{Check if pellatParams object is valid}
\usage{
isValidPellatParams(object)
}
\arguments{
\item{object}{A pellatParams object}
}
\value{
TRUE if valid, FALSE otherwise
}
\description{
Simple boolean check for object validity.
}
\examples{
pt = pellatParams(c(msy = 500, fmsy = 0.3, virgin = 10000))
isValidPellatParams(pt)
}
