% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jabba-methods.R
\name{jabba2biodyn}
\alias{jabba2biodyn}
\title{Coerce a JABBA fit to an mpb \code{biodyn}}
\usage{
jabba2biodyn(object)
}
\arguments{
\item{object}{A JABBA fit (or list with a \code{fit} element).}
}
\value{
An mpb \code{biodyn}.
}
\description{
Copies catch, biomass, index, median \code{r}, \(K\), shape and
initial depletion from a JABBA fit into an mpb \code{biodyn}.
Accepts a JABBA fit or a wrapper with \code{fit$timeseries}.
}
\examples{
\dontrun{
fit <- runJABBA(ple4, ple4brp, index = ebiomass)
bd  <- jabba2biodyn(fit)
}

}
\seealso{
\code{\link{runJABBA}}
}
