% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pMetricsFuncs.R
\name{kobeLines}
\alias{kobeLines}
\alias{lns}
\title{Create Kobe Plot Reference Lines}
\usage{
kobeLines(xlim = c(0, 2), ylim = c(0, 2))

lns(...)
}
\arguments{
\item{xlim}{Numeric vector of length 2 specifying x-axis limits (default = c(0, 2))}

\item{ylim}{Numeric vector of length 2 specifying y-axis limits (default = c(0, 2))}
}
\value{
A list containing:
  \item{vline}{Vertical line at x = 1 (B/BMSY = 1)}
  \item{hline}{Horizontal line at y = 1 (F/FMSY = 1)}
  \item{data}{Data frame with line coordinates for plotting}
}
\description{
Creates reference lines for Kobe plots (B/BMSY = 1 and F/FMSY = 1).
These lines divide the Kobe plot into quadrants.
}
\details{
These reference lines are used to divide Kobe plots into four quadrants:
- Green: B >= 1, F <= 1 (sustainable)
- Red: B < 1, F > 1 (collapsed)
- Yellow: B < 1, F <= 1 (overfished only)
- Orange: B >= 1, F > 1 (overfishing only)
}
\examples{
\dontrun{
lns <- kobeLines()
# Use in ggplot:
# ggplot() + 
#   geom_vline(data = lns$data, aes(xintercept = x), linetype = 2) +
#   geom_hline(data = lns$data, aes(yintercept = y), linetype = 2)
}

}
