% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag.R
\docType{methods}
\name{lag,FLBRP-method}
\alias{lag,FLBRP-method}
\title{Estimate population dynamics lag}
\usage{
\S4method{lag}{FLBRP}(object, ...)
}
\arguments{
\item{object}{An object of class `FLBRP`}

\item{...}{Additional arguments (not currently used)}
}
\value{
An FLQuant object with 1 quant dimension, containing the age indices
where SSB contribution is maximized, indicating the population dynamics lag
}
\description{
This method estimates the lag in population dynamics by finding the age with maximum SSB contribution.
Excludes the plusgroup from the calculation.
}
\examples{
data(ple4brp)

# Estimate lag using SSB age distribution
lag_time=lag(ple4brp)
}
\seealso{
[FLBRP]
}
