% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mFn.R
\name{M2Fn}
\alias{M2Fn}
\alias{M2Fn,FLQuant-method}
\alias{M2Fn,FLStock-method}
\alias{M2Fn,FLBRP-method}
\title{Calculate Predation/Variable Mortality (M2)}
\usage{
M2Fn(x, y = 0.025)

\S4method{M2Fn}{FLQuant}(x, y = 0.025)

\S4method{M2Fn}{FLStock}(x, y = 0.025)

\S4method{M2Fn}{FLBRP}(x, y = 0.025)
}
\arguments{
\item{x}{An FLQuant object containing natural mortality, or an object with 
natural mortality method (e.g., FLStock, FLBRP)}

\item{y}{Optional numeric value for M1 (default = 0.025)}
}
\value{
An FLQuant object with M2 mortality rates (M - M1)
}
\description{
Calculates the variable/predation component of natural mortality (M2)
as the difference between total natural mortality (M) and background
mortality (M1).
}
\details{
M2 represents the variable component of natural mortality, often associated
with predation or other density-dependent factors. It is calculated as:
M2 = M - M1
where M is total natural mortality and M1 is the background mortality.
}
\examples{
\dontrun{
library(FLCore)
library(FLBRP)
data(ple4brp)
# Using FLBRP object
M2=M2Fn(ple4brp)

# Using FLQuant object
m_vals=m(ple4brp)
M2=M2Fn(m_vals)
}

}
