% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refpt-processing.R
\name{mergeTimeSeriesRefpts}
\alias{mergeTimeSeriesRefpts}
\title{Merge Time Series and Reference Point Data}
\usage{
mergeTimeSeriesRefpts(ts_data, refpt_data, merge_by = c(".id", "year"))
}
\arguments{
\item{ts_data}{Time series data}

\item{refpt_data}{Reference point data}

\item{merge_by}{Character vector of column names to merge by}
}
\value{
Merged operating model data
}
\description{
Merge time series data with reference point data for operating model analysis.
}
