% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel-utils.R
\name{monitorProgress}
\alias{monitorProgress}
\title{Monitor Parallel Processing Progress}
\usage{
monitorProgress(total_tasks, completed_tasks, start_time)
}
\arguments{
\item{total_tasks}{Total number of tasks}

\item{completed_tasks}{Number of completed tasks}

\item{start_time}{Start time of processing}
}
\value{
Progress information
}
\description{
Monitor progress of parallel processing tasks.
}
