% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pella_tomlinson.R
\name{msy,FLPar-method}
\alias{msy,FLPar-method}
\title{Method for MSY (FLPar)}
\usage{
\S4method{msy}{FLPar}(params)
}
\arguments{
\item{params}{FLPar object.}
}
\value{
Numeric value or FLQuant object.
}
\description{
Method for calculating MSY when parameters are provided as an FLPar object.
}
