% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pella_tomlinson.R
\name{msy,numeric-method}
\alias{msy,numeric-method}
\title{Method for MSY (numeric)}
\usage{
\S4method{msy}{numeric}(params)
}
\arguments{
\item{params}{Numeric vector.}
}
\value{
Numeric value.
}
\description{
Method for calculating MSY when parameters are provided as a numeric vector.
}
