% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pella_tomlinson.R
\name{msy,PellaTomlinson-method}
\alias{msy,PellaTomlinson-method}
\title{Method for MSY (PellaTomlinson)}
\usage{
\S4method{msy}{PellaTomlinson}(params)
}
\arguments{
\item{params}{PellaTomlinson object.}
}
\value{
Numeric value or FLQuant object.
}
\description{
Method for calculating MSY when the PellaTomlinson object is used.
}
