% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pella_tomlinson.R
\name{msy}
\alias{msy}
\title{Calculate Maximum Sustainable Yield (MSY)}
\usage{
msy(params, ...)

msy(params, ...)
}
\arguments{
\item{params}{FLPar object.}

\item{...}{Additional arguments.}
}
\value{
Numeric value of MSY.

Numeric value or FLQuant object.
}
\description{
Calculates the Maximum Sustainable Yield (MSY) given model parameters.

Generic function for calculating Maximum Sustainable Yield (MSY).
}
\examples{
params = c(r = 0.5, p = 0.25, virgin = 1.0)
msy_val = msy(params)

}
