% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pella_tomlinson.R
\name{numeric-methods}
\alias{numeric-methods}
\title{Methods for numeric params}
\description{
Methods for calculating production, MSY, BMSY, FMSY, p, r, param
when parameters are provided as numeric vectors.
}
\keyword{internal}
