% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pella_tomlinson.R
\name{p,FLPar-method}
\alias{p,FLPar-method}
\title{Method for p (FLPar)}
\usage{
\S4method{p}{FLPar}(params)
}
\arguments{
\item{params}{FLPar object.}
}
\value{
FLPar object with p parameter.
}
\description{
Method for calculating the shape parameter p when parameters are provided as an FLPar object.
}
