% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pella_tomlinson.R
\name{p,numeric-method}
\alias{p,numeric-method}
\title{Method for p (numeric)}
\usage{
\S4method{p}{numeric}(params)
}
\arguments{
\item{params}{Numeric vector.}
}
\value{
Numeric value of the shape parameter p.
}
\description{
Method for calculating the shape parameter p when parameters are provided as a numeric vector.
}
