% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pella_tomlinson.R
\name{p,PellaTomlinson-method}
\alias{p,PellaTomlinson-method}
\title{p method for PellaTomlinson class - derives p from bmsy and virgin}
\usage{
\S4method{p}{PellaTomlinson}(params)
}
\arguments{
\item{params}{PellaTomlinson object.}
}
\value{
FLPar object with p parameter.
}
\description{
Derives the shape parameter p from BMSY and virgin biomass.
}
