% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pella_tomlinson.R
\name{p}
\alias{p}
\title{Generic function for p}
\usage{
p(params, ...)
}
\arguments{
\item{params}{FLPar object.}

\item{...}{Additional arguments.}
}
\value{
FLPar object with p parameter.
}
\description{
Generic function for calculating the shape parameter p.
}
