% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pellatUtils.R
\name{p2shape}
\alias{p2shape}
\title{Convert shape parameter p to BMSY/K ratio}
\usage{
p2shape(p)
}
\arguments{
\item{p}{Shape parameter}
}
\value{
BMSY/K ratio
}
\description{
Convert shape parameter p to BMSY/K ratio
}
