% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebuild-analysis.R
\name{parallelRebuildAnalysis}
\alias{parallelRebuildAnalysis}
\title{Parallel Rebuild Analysis for Multiple Stocks}
\usage{
parallelRebuildAnalysis(
  flbrp_list,
  stock_names,
  ncores = NULL,
  rebuild_args = list()
)
}
\arguments{
\item{flbrp_list}{List of FLBRP objects}

\item{stock_names}{Character vector of stock names}

\item{ncores}{Number of cores to use (default=detectCores()-4)}

\item{rebuild_args}{List of arguments to pass to rebuild function}
}
\value{
List of rebuild results
}
\description{
Perform parallel rebuild analysis for multiple stocks.
}
