% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebuild-analysis.R
\name{parallelRefptExtraction}
\alias{parallelRefptExtraction}
\title{Parallel Reference Point Extraction}
\usage{
parallelRefptExtraction(
  flbrp_list,
  ncores = NULL,
  refpts_to_extract = c("msy", "f0.1", "spr.30")
)
}
\arguments{
\item{flbrp_list}{List of FLBRP objects}

\item{ncores}{Number of cores to use (default=detectCores()-4)}

\item{refpts_to_extract}{Character vector of reference points to extract}
}
\value{
Extracted reference points
}
\description{
Extract reference points from FLBRP objects using parallel processing.
}
