% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pella_tomlinson.R
\name{param,FLPar-method}
\alias{param,FLPar-method}
\title{Method for param (FLPar)}
\usage{
\S4method{param}{FLPar}(params)
}
\arguments{
\item{params}{FLPar object.}
}
\value{
FLPar object with estimated parameters.
}
\description{
Method for estimating both r and p parameters when parameters are provided as an FLPar object.
}
