% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pella_tomlinson.R
\name{param,numeric-method}
\alias{param,numeric-method}
\title{Method for param (numeric)}
\usage{
\S4method{param}{numeric}(params)
}
\arguments{
\item{params}{Numeric vector.}
}
\value{
Numeric vector with estimated parameters.
}
\description{
Method for estimating both r and p parameters when parameters are provided as a numeric vector.
}
