% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pella_tomlinson.R
\name{param,PellaTomlinson-method}
\alias{param,PellaTomlinson-method}
\title{Method for param (PellaTomlinson)}
\usage{
\S4method{param}{PellaTomlinson}(params)
}
\arguments{
\item{params}{PellaTomlinson object.}
}
\value{
FLPar object with estimated parameters.
}
\description{
Method for estimating both r and p parameters when parameters are provided as a PellaTomlinson object.
}
