% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pella_tomlinson.R
\name{param}
\alias{param}
\title{Estimate Both r and p Parameters}
\usage{
param(params, ...)

param(params, ...)
}
\arguments{
\item{params}{FLPar object.}

\item{...}{Additional arguments.}
}
\value{
Updated parameter vector with estimated 'r' and 'p' values

FLPar object with estimated parameters.
}
\description{
Estimates both r and p parameters from MSY, BMSY, and virgin biomass.

Generic function for estimating both r and p parameters.
}
\examples{
params = c(msy = 0.1, bmsy = 0.7, virgin = 1.0)
estimated_params = param(params)

}
