% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pella_tomlinson.R
\name{params}
\alias{params}
\alias{params,PellaTomlinson-method}
\title{Accessor method for parameters}
\usage{
params(object, ...)

\S4method{params}{PellaTomlinson}(object)
}
\arguments{
\item{object}{Pella-Tomlinson object}

\item{...}{Additional arguments}
}
\value{
FLPar object containing the model parameters
}
\description{
Retrieves the parameters from a Pella-Tomlinson object.
}
\examples{
params = FLPar(r = 0.5, p = 0.25, virgin = 1.0, bmsy = 0.7, msy = 0.1)
pt_model = PellaTomlinson(params)
model_params = params(pt_model)

}
