% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pellat.R
\name{pellat}
\alias{pellat}
\alias{pellat,FLPar-method}
\alias{pellat,FLBRP-method}
\alias{pellat,FLBRPs-method}
\title{Pella-Tomlinson Parameter Calculator}
\usage{
pellat(object, biomass, ...)

\S4method{pellat}{FLPar}(object)

\S4method{pellat}{FLBRP}(object, biomass = "ssb", ...)

\S4method{pellat}{FLBRPs}(object, biomass = "ssb", ...)
}
\arguments{
\item{object}{FLPar object containing parameters}

\item{biomass}{Character string specifying biomass type (optional)}

\item{...}{Additional arguments}
}
\value{
FLPar object with calculated parameters
}
\description{
Main function to calculate Pella-Tomlinson parameters from various input combinations.
This is the primary function for parameter calculation.
}
\examples{
params = FLPar(bmsy = 500, k = 1000, fmsy = 0.2)
result = pellat(params)
}
