% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pellat.R
\name{pellaTMleCore}
\alias{pellaTMleCore}
\title{Core MLE function for Pella-Tomlinson parameter estimation}
\usage{
pellaTMleCore(object, biomass = "ssb", ...)
}
\description{
Core MLE function for Pella-Tomlinson parameter estimation
}
\keyword{internal}
