% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pella_tomlinson.R
\docType{class}
\name{PellaTomlinson-class}
\alias{PellaTomlinson-class}
\title{Pella-Tomlinson Production Model Class}
\description{
S4 class representing a Pella-Tomlinson production model with parameters
and stock biomass data.
}
\section{Slots}{

\describe{
\item{\code{params}}{FLPar object containing model parameters (r, p, virgin, bmsy, msy)}

\item{\code{stock}}{FLQuant object containing stock biomass data}
}}

