% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pella_tomlinson.R
\name{PellaTomlinson-methods}
\alias{PellaTomlinson-methods}
\title{S4 Method Definitions for the PellaTomlinson class}
\description{
Generic functions for the PellaTomlinson class.

Methods for calculating production, MSY, BMSY, FMSY, p, r, param
when the PellaTomlinson object is used as the method's first argument.
}
\details{
These functions are used to define the methods for the S4 class.
}
\keyword{internal}
