% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pella_tomlinson.R
\name{plot,PellaTomlinson,missing-method}
\alias{plot,PellaTomlinson,missing-method}
\title{Plot method for PellaTomlinson class using ggplot}
\usage{
\S4method{plot}{PellaTomlinson,missing}(x, y, ...)
}
\arguments{
\item{x}{PellaTomlinson object.}

\item{y}{Missing (production is plotted against biomass).}

\item{...}{Additional arguments for ggplot.}
}
\description{
Plots the production curve of the PellaTomlinson object.
}
