% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pellatParams.R
\name{plot,pellatParams,ANY-method}
\alias{plot,pellatParams,ANY-method}
\title{Plot method for pellatParams}
\usage{
\S4method{plot}{pellatParams,ANY}(x, y, ...)
}
\arguments{
\item{x}{A pellatParams object}

\item{y}{Not used (required for S4 method signature)}

\item{...}{Additional arguments passed to ggplot2 functions}
}
\value{
A ggplot2 object
}
\description{
Creates a ggplot2 plot showing the production function and reference points.
}
\examples{
pt = pellatParams(c(msy = 500, fmsy = 0.3, virgin = 10000))
pt = estimateParams(pt)
plot(pt)

# Save plot to file
p = plot(pt)
ggplot2::ggsave("production_function.png", p, width = 10, height = 6)
}
