% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-plots.R
\name{plotOMPerformance}
\alias{plotOMPerformance}
\title{Plot Operating Model Performance}
\usage{
plotOMPerformance(
  performance_data,
  refpts,
  plot_type = "scatter",
  metric = "bias",
  facet_by = NULL
)
}
\arguments{
\item{performance_data}{Performance data}

\item{refpts}{Reference points to evaluate}

\item{plot_type}{Type of plot ("scatter", "heatmap", "bar")}

\item{metric}{Performance metric to plot}

\item{facet_by}{Variable to facet by (optional)}
}
\value{
ggplot object
}
\description{
Create operating model performance plots.
}
