% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-plots.R
\name{plotRebuildTimes}
\alias{plotRebuildTimes}
\title{Plot Rebuild Times}
\usage{
plotRebuildTimes(
  rebuild_data,
  col_names,
  plot_type = "histogram",
  x_var = NULL,
  facet_by = NULL
)
}
\arguments{
\item{rebuild_data}{Rebuild time data}

\item{col_names}{Column names for plotting}

\item{plot_type}{Type of plot ("histogram", "density", "boxplot", "scatter")}

\item{x_var}{Variable for x-axis (for scatter plots)}

\item{facet_by}{Variable to facet by (optional)}
}
\value{
ggplot object
}
\description{
Create rebuild time analysis plots.
}
