% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotRebuildTrajectories}
\alias{plotRebuildTrajectories}
\title{Plot rebuilding trajectories}
\usage{
plotRebuildTrajectories(
  rebuild_data,
  target_line = 1,
  title = "Rebuilding Trajectories"
)
}
\arguments{
\item{rebuild_data}{Data frame with rebuilding trajectories}

\item{target_line}{Target biomass level (default = 1 for BMSY)}

\item{title}{Plot title}
}
\value{
ggplot object
}
\description{
Create a ggplot of rebuilding trajectories
}
