% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-plots.R
\name{plotRefptCorrelations}
\alias{plotRefptCorrelations}
\title{Plot Reference Point Correlations}
\usage{
plotRefptCorrelations(
  refpt_data,
  refpts,
  plot_type = "correlation",
  method = "pearson"
)
}
\arguments{
\item{refpt_data}{Reference point data}

\item{refpts}{Reference points to include in correlation analysis}

\item{plot_type}{Type of plot ("correlation", "scatter_matrix", "heatmap")}

\item{method}{Correlation method ("pearson", "spearman", "kendall")}
}
\value{
ggplot object or list of plots
}
\description{
Create correlation plots between reference points.
}
