% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pella_tomlinson.R
\name{print,PellaTomlinson-method}
\alias{print,PellaTomlinson-method}
\title{Print method for PellaTomlinson class}
\usage{
\S4method{print}{PellaTomlinson}(x, ...)
}
\arguments{
\item{x}{PellaTomlinson object.}

\item{...}{Additional arguments.}
}
\description{
Prints a summary of the PellaTomlinson object.
}
