% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pellatParams.R
\name{print,pellatParams-method}
\alias{print,pellatParams-method}
\title{Print method for pellatParams}
\usage{
\S4method{print}{pellatParams}(x, ...)
}
\arguments{
\item{x}{A pellatParams object}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns the object
}
\description{
Print method for pellatParams
}
\examples{
pt = pellatParams(c(msy = 500, fmsy = 0.3, virgin = 10000))
print(pt)
}
