% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pellatParams.R
\name{print.summary.pellatParams}
\alias{print.summary.pellatParams}
\title{Print method for summary.pellatParams}
\usage{
\method{print}{summary.pellatParams}(x, ...)
}
\arguments{
\item{x}{A summary.pellatParams object}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns the object
}
\description{
Print method for summary.pellatParams
}
