% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobe.R
\name{probFn2}
\alias{probFn2}
\title{Calculate Kobe Quadrant Probabilities (HPD Region Method)}
\usage{
probFn2(
  x,
  y,
  prob = c(0.5, 0.75, 0.95),
  n = 21,
  h = c(bandwidth.nrd(x), bandwidth.nrd(y)),
  lims = NULL,
  na.rm = FALSE
)
}
\arguments{
\item{x}{A numeric vector}

\item{y}{A numeric vector}

\item{prob}{Numeric vector of probability levels (default = c(0.5, 0.75, 0.95))}

\item{n}{Number of points at which to evaluate the density grid (default = 21)}

\item{h}{Bandwidth of 2D kernel smoother. If not specified, defaults to
bandwidth.nrd() values for x and y}

\item{lims}{Limits, specified as (x.lower, x.upper, y.lower, y.upper)
(passed to kde2d)}

\item{na.rm}{Logical; if TRUE, any NA and NaN's are removed from x and y
before calculations (default = FALSE)}
}
\value{
A data.frame with three variables:
  \item{x}{x coordinates of the grid points, vector of length n}
  \item{y}{y coordinates of the grid points, vector of length n}
  \item{level}{Contours corresponding to prob}
}
\description{
Calculates the probability of observations occurring in a 2D cell using
HPDregionplot. Given a sample, calculates the bivariate region of highest
marginal posterior density for two variables, using kde2d from MASS to
calculate a bivariate density.
}
\details{
This function uses HPDregionplot from the coda package to calculate
highest posterior density regions for bivariate data. It's an alternative
implementation to the simpler probFn function which uses discrete thresholds.
}
\examples{
\dontrun{
x <- rnorm(20)
y <- rnorm(20)
probFn2(x, y)
}

}
\keyword{internal}
