% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refpt-processing.R
\name{processBreakpointData}
\alias{processBreakpointData}
\title{Process Breakpoint Data for Multiple Scenarios}
\usage{
processBreakpointData(flbrp_list, sr_params, scenarios, srrs)
}
\arguments{
\item{flbrp_list}{List of FLBRP objects}

\item{sr_params}{Data frame with stock-recruitment parameters}

\item{scenarios}{Character vector of scenario names}

\item{srrs}{Character vector of stock-recruitment relationship names}
}
\value{
Processed breakpoint data
}
\description{
Process breakpoint data for multiple scenarios and stock-recruitment relationships.
}
