% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-generic.R, R/flbrp_process_error.R
\name{processError}
\alias{processError}
\alias{processError,FLBRP-method}
\title{Calculate process error}
\usage{
processError(object, ...)

\S4method{processError}{FLBRP}(object)
}
\arguments{
\item{object}{An FLBRP object.}

\item{...}{Additional arguments}
}
\value{
An FLQuants object containing process error metrics.
}
\description{
This function calculates process error for an FLBRP object.
}
\examples{
\dontrun{
library(FLCore)
library(FLBRP)
# Create an FLBRP object
eq <- lhEql(lhPar(FLPar(linf=250, s=0.9)))
# Calculate process error
pe <- processError(eq)
}

}
