% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pella_tomlinson.R
\name{production,FLQuant,FLPar-method}
\alias{production,FLQuant,FLPar-method}
\title{Method for production (FLPar)}
\usage{
\S4method{production}{FLQuant,FLPar}(biomass, params)
}
\arguments{
\item{biomass}{FLQuant object.}

\item{params}{FLPar object.}
}
\value{
FLQuant object.
}
\description{
Method for calculating production when parameters are provided as an FLPar object.
}
