% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pella_tomlinson.R
\name{production,FLQuant,PellaTomlinson-method}
\alias{production,FLQuant,PellaTomlinson-method}
\title{Method for production (PellaTomlinson)}
\usage{
\S4method{production}{FLQuant,PellaTomlinson}(biomass, params)
}
\arguments{
\item{biomass}{FLQuant object.}

\item{params}{PellaTomlinson object.}
}
\value{
FLQuant object.
}
\description{
Method for calculating production when the PellaTomlinson object is used.
}
