% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pella_tomlinson.R
\name{production,numeric,numeric-method}
\alias{production,numeric,numeric-method}
\title{Method for production (numeric)}
\usage{
\S4method{production}{numeric,numeric}(biomass, params)
}
\arguments{
\item{biomass}{Numeric vector.}

\item{params}{Numeric vector.}
}
\value{
Numeric vector.
}
\description{
Method for calculating production when parameters are provided as a numeric vector.
}
