% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pella_tomlinson.R
\name{production,PellaTomlinson,missing-method}
\alias{production,PellaTomlinson,missing-method}
\title{Method for production (PellaTomlinson)}
\usage{
\S4method{production}{PellaTomlinson,missing}(biomass, params)
}
\arguments{
\item{biomass}{Missing (object's stock data is used).}

\item{params}{PellaTomlinson object.}
}
\value{
FLQuant object.
}
\description{
Method for calculating production when the PellaTomlinson object is used.
}
