% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pellat.R
\name{production}
\alias{production}
\alias{production,FLPar,FLQuant-method}
\title{Calculate Pella-Tomlinson Production}
\usage{
production(params, biomass, ...)

\S4method{production}{FLPar,FLQuant}(params, biomass)
}
\arguments{
\item{params}{FLPar object with r, p, virgin parameters}

\item{biomass}{FLQuant object with biomass values}

\item{...}{Additional arguments}
}
\value{
FLQuant object with production values
}
\description{
Calculate production (surplus production) for given Pella-Tomlinson parameters and biomass.
This is the main function for calculating production in the Pella-Tomlinson surplus production model.
}
\examples{
params = FLPar(r = 0.5, p = 1, virgin = 1000)
biomass = FLQuant(500)
result = production(params, biomass)
}
