% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pellat.R
\name{ptMle}
\alias{ptMle}
\alias{ptMle,data.frame-method}
\alias{ptMle,FLBRP-method}
\alias{ptMle,FLBRPs-method}
\alias{ptMle,list-method}
\title{Pella-Tomlinson Maximum Likelihood Estimation}
\usage{
ptMle(object, biomass = "ssb", ...)

\S4method{ptMle}{data.frame}(object, biomass = "ssb", ...)

\S4method{ptMle}{FLBRP}(object, biomass = "ssb", ...)

\S4method{ptMle}{FLBRPs}(object, biomass = "ssb", ...)

\S4method{ptMle}{list}(object, biomass = "ssb", ...)
}
\arguments{
\item{object}{The object to fit the model to:
- data.frame: Must contain biomass and yield columns
- FLBRP: Single FLBRP object
- FLBRPs: Collection of FLBRP objects
- list: List of FLBRP objects}

\item{biomass}{Character string specifying the biomass column name (default: "ssb")}

\item{...}{Additional arguments passed to optimization}
}
\value{
For single objects: FLPar object with estimated parameters
        For collections: List of FLPar objects
}
\description{
Fits Pella-Tomlinson production model using maximum likelihood estimation.
Dispatched on different object types: data.frame, FLBRP, FLBRPs, and list.
}
\examples{
# data.frame method
df = data.frame(ssb = c(1000, 2000, 3000, 4000, 5000),
                yield = c(100, 200, 250, 200, 150))
result = ptMle(df, biomass = "ssb")
}
