% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pella_tomlinson.R
\name{r,FLPar-method}
\alias{r,FLPar-method}
\title{Method for r (FLPar)}
\usage{
\S4method{r}{FLPar}(params)
}
\arguments{
\item{params}{FLPar object.}
}
\value{
FLPar object with r parameter.
}
\description{
Method for calculating the intrinsic growth rate r when parameters are provided as an FLPar object.
}
