% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pella_tomlinson.R
\name{r,numeric-method}
\alias{r,numeric-method}
\title{Method for r (numeric)}
\usage{
\S4method{r}{numeric}(params)
}
\arguments{
\item{params}{Numeric vector.}
}
\value{
Numeric value of the intrinsic growth rate r.
}
\description{
Method for calculating the intrinsic growth rate r when parameters are provided as a numeric vector.
}
