% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pella_tomlinson.R
\name{r,PellaTomlinson-method}
\alias{r,PellaTomlinson-method}
\title{r method for PellaTomlinson class - derives r from msy, bmsy, and virgin}
\usage{
\S4method{r}{PellaTomlinson}(params)
}
\arguments{
\item{params}{PellaTomlinson object.}
}
\value{
FLPar object with r parameter.
}
\description{
Derives the intrinsic growth rate r from MSY, BMSY, and virgin biomass.
}
