% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pella_tomlinson.R
\name{r}
\alias{r}
\title{Generic function for r}
\usage{
r(params, ...)
}
\arguments{
\item{params}{FLPar object.}

\item{...}{Additional arguments.}
}
\value{
FLPar object with r parameter.
}
\description{
Generic function for calculating the intrinsic growth rate r.
}
