% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rboot.R
\name{rboot_years}
\alias{rboot_years}
\title{Resample FLQuant values from specified years}
\usage{
rboot_years(object, years, n = 100)
}
\arguments{
\item{object}{An FLQuant object}

\item{years}{A vector of years to sample from (must be in dimnames(object)$year)}

\item{n}{Number of resampled iterations to return}
}
\value{
An FLQuant with resampled values in the iter dimension
}
\description{
Resample FLQuant values from specified years
}
