% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodyn-rebuild.R
\docType{class}
\name{Rebuild-class}
\alias{Rebuild-class}
\title{Rebuild Class}
\description{
Rebuild Class
}
\section{Slots}{

\describe{
\item{\code{params}}{An \code{FLPar} object containing model parameters}

\item{\code{nyrs}}{Numeric, number of years for projection}

\item{\code{niters}}{Numeric, number of iterations}
}}

