% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebuild_multi.R
\name{rebuild2}
\alias{rebuild2}
\title{Rebuild trajectories for multiple scenarios (version 2)}
\usage{
rebuild2(
  object,
  targetF = computeRefpts(object)["msy", "harvest"] * 0,
  targetSSB = computeRefpts(object)["msy", "ssb"],
  initial = 100,
  minVal = 1e-06,
  maxVal = 1,
  growthRate = 0.25
)
}
\arguments{
\item{object}{An FLBRP object}

\item{targetF}{Target fishing mortality (default is 0)}

\item{targetSSB}{Target SSB (default is MSY SSB)}

\item{initial}{Number of initial depletion levels to evaluate (default 100)}

\item{minVal}{Minimum depletion level (default 1e-6)}

\item{maxVal}{Maximum depletion level (default 1)}

\item{growthRate}{Growth rate parameter for SSB sequence (default 0.25)}
}
\value{
FLStock object with rebuilding trajectories
}
\description{
Projects rebuilding trajectories from different initial SSB levels using version 2 algorithm
}
