% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLBRP-rebuild.R
\name{rebuildTime,FLStock-method}
\alias{rebuildTime,FLStock-method}
\title{Calculate rebuilding time}
\usage{
\S4method{rebuildTime}{FLStock}(object, nx = 101)
}
\arguments{
\item{object}{An FLStock object}

\item{nx}{Number of interpolation points (default = 101)}
}
\value{
A data frame with columns year and initial
}
\description{
Calculate rebuilding time
}
