% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/rebuild-methods.R
\name{rebuildTime}
\alias{rebuildTime}
\alias{rebuildTime,biodyn-method}
\alias{rebuildTime,FLStock-method}
\title{Calculate rebuilding time}
\usage{
rebuildTime(object, ...)

\S4method{rebuildTime}{biodyn}(
  object,
  target = refpts(object)["bmsy"],
  nInitial = 100,
  growthRate = 0.3,
  minVal = 1e-06,
  maxVal = 1,
  nx = 101
)

\S4method{rebuildTime}{FLStock}(object, nx = 101)
}
\arguments{
\item{object}{An object containing rebuilding trajectories}

\item{...}{Additional arguments}
}
\value{
A data frame with columns year and initial
}
\description{
Calculate rebuilding time
}
