% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-generic.R, R/rebuild_biodyn.R
\name{rebuildTime2}
\alias{rebuildTime2}
\alias{rebuildTime2,FLPar-method}
\title{Calculate Rebuild Time using FLPar parameters (version 2)}
\usage{
rebuildTime2(object, ...)

\S4method{rebuildTime2}{FLPar}(object, nyrs = 50, ftar = 0, timing = 0, minVal = 0.05, ...)
}
\arguments{
\item{object}{An FLPar object containing parameters (r, k, p, B0)}

\item{...}{Additional arguments}

\item{nyrs}{Number of years for projection (default = 50)}

\item{minVal}{Minimum depletion value as proportion of BMSY (default = 0.05)}
}
\value{
A data frame with rebuild times

A data frame with columns:
  \item{initial}{Initial depletion level relative to BMSY}
  \item{T_recover}{Time to recover to BMSY}
}
\description{
A simplified and faster version of rebuildTime that works directly with FLPar objects.
This function calculates rebuilding time using Pella-Tomlinson dynamics parameters.
}
\examples{
# Create FLPar with Pella-Tomlinson parameters
params=FLPar(r = 0.5, k = 1000, p = 1, B0 = 1000)
rebuild_time=rebuildTime2(params)

# With custom parameters
rebuild_time=rebuildTime2(params, nyrs = 100, minVal = 0.1)
}
