% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-generic.R
\name{rebuildTime3}
\alias{rebuildTime3}
\title{Calculate Rebuild Time using FLPar parameters (version 3)}
\usage{
rebuildTime3(object, ...)
}
\arguments{
\item{object}{An object with parameters}

\item{...}{Additional arguments}
}
\value{
A data frame with rebuild times
}
\description{
Calculate Rebuild Time using FLPar parameters (version 3)
}
