% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr-rs.R
\name{refCreate}
\alias{refCreate}
\alias{refCreate,FLBRP-method}
\title{Create reference point FLPar object}
\usage{
refCreate(object, ref, value = NA, quant = ref, ...)

\S4method{refCreate}{FLBRP}(object, ref, value = NA, quant = ref, ...)
}
\arguments{
\item{object}{An object (typically FLBRP)}

\item{ref}{Reference point name (character)}

\item{value}{Value for the reference point (numeric, default NA)}

\item{quant}{Quant to set (default is ref)}

\item{...}{Additional arguments}
}
\value{
FLPar object with reference point structure

An FLPar object with reference point structure.
}
\description{
Create reference point FLPar object
}
\details{
For an FLBRP object, returns a new FLPar reference point object with the specified values. The input FLBRP object is not modified.
For an FLPar object, returns a new FLPar reference point object mapping parameter names to quant columns. The input FLPar object is not modified.
}
