% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refptsEB.R
\name{refptsEB}
\alias{refptsEB}
\alias{refptsEB,FLBRP-method}
\title{Calculate Reference Points Including ebiomass (EB)}
\usage{
refptsEB(object, ...)

\S4method{refptsEB}{FLBRP}(object, biomass = "ssb", ...)
}
\arguments{
\item{object}{An FLBRP object.}
}
\value{
An FLPar object with reference points, including the EB quant.
}
\description{
Calculates reference points for an FLBRP object, including the ebiomass (EB) as an additional quant.
}
