% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-generic.R, R/analysis.R
\name{results}
\alias{results}
\alias{results,FLStock,FLBRP-method}
\alias{results,list,FLBRP-method}
\title{Calculate Performance Metrics from Operating Model Results}
\usage{
results(object, eq, ...)

\S4method{results}{FLStock,FLBRP}(object, eq, ftar = 0.5)

\S4method{results}{list,FLBRP}(object, eq, ftar = 0.5)
}
\arguments{
\item{object}{An FLStock object or list of FLStock objects (operating model results)}

\item{eq}{An FLBRP object containing reference points}

\item{...}{Additional arguments}
}
\value{
An FLQuants object containing relative metrics
}
\description{
Calculates a suite of performance metrics from operating model (OM) results
relative to reference points from an equilibrium (eq) object.
}
