% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pellatUtils.R
\name{rFmsy2p}
\alias{rFmsy2p}
\title{Calculate shape parameter p from r and FMSY}
\usage{
rFmsy2p(r, Fmsy)
}
\arguments{
\item{r}{Intrinsic growth rate}

\item{Fmsy}{Fishing mortality at MSY}
}
\value{
Shape parameter p
}
\description{
Calculate shape parameter p from r and FMSY
}
