% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pella_tomlinson.R
\name{rFn}
\alias{rFn}
\title{Calculate Intrinsic Growth Rate r from MSY, BMSY, and Virgin Biomass}
\usage{
rFn(params)
}
\arguments{
\item{params}{Named numeric vector containing 'msy', 'bmsy', and 'virgin' parameters}
}
\value{
Numeric value of the intrinsic growth rate r
}
\description{
Calculates the intrinsic growth rate r given MSY, BMSY, and virgin biomass.
}
\details{
First calculates p using pFn(), then solves for r using the relationship:
MSY = (r/p) * BMSY * (1 - (BMSY/virgin)^p)
}
\examples{
params = c(msy = 0.1, bmsy = 0.7, virgin = 1.0)
r_val = rFn(params)

}
