% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr-rs.R
\name{rickerInv}
\alias{rickerInv}
\title{Inverse Ricker stock-recruitment relationship}
\usage{
rickerInv(params, rec)
}
\arguments{
\item{params}{FLPar with parameters 'a' and 'b'}

\item{rec}{Recruitment values}
}
\value{
SSB values
}
\description{
Inverse Ricker stock-recruitment relationship
}
\keyword{internal}
