% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr-rs.R
\name{rickerSSBMaxRec}
\alias{rickerSSBMaxRec}
\title{Maximum recruitment SSB for Ricker model}
\usage{
rickerSSBMaxRec(b)
}
\arguments{
\item{b}{Ricker parameter b}
}
\value{
SSB at maximum recruitment
}
\description{
Maximum recruitment SSB for Ricker model
}
\keyword{internal}
