% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr-rs.R
\name{rmax}
\alias{rmax}
\alias{rmax,FLBRP,missing-method}
\alias{rmax,FLBRP,numeric-method}
\title{Maximum recruitment for a stock-recruitment model}
\usage{
rmax(object, ratio, ...)

rmax(object, ratio, ...)

\S4method{rmax}{FLBRP,missing}(object, ratio, ...)

\S4method{rmax}{FLBRP,numeric}(object, ratio, ...)
}
\arguments{
\item{object}{An object (typically FLBRP)}

\item{ratio}{Optional numeric ratio (default missing)}

\item{...}{Additional arguments}
}
\value{
Maximum recruitment (numeric or FLPar)

Maximum recruitment (numeric or FLPar)
}
\description{
Maximum recruitment for a stock-recruitment model

Maximum recruitment for a stock-recruitment model
}
