% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr-rs.R
\name{rmsy}
\alias{rmsy}
\alias{rmsy,FLBRP,missing-method}
\alias{rmsy,FLBRP,numeric-method}
\title{Recruitment at MSY for a stock-recruitment model}
\usage{
rmsy(object, ratio = 1, ...)

\S4method{rmsy}{FLBRP,missing}(object, ratio = 1, ...)

\S4method{rmsy}{FLBRP,numeric}(object, ratio = 1, ...)
}
\arguments{
\item{object}{An object (typically FLBRP)}

\item{ratio}{Optional numeric ratio (default 1.0)}

\item{...}{Additional arguments}
}
\value{
Recruitment at MSY (numeric or FLPar)
}
\description{
Recruitment at MSY for a stock-recruitment model
}
