% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebuild-pellaT.R
\name{rTime}
\alias{rTime}
\alias{rTime,numeric-method}
\alias{rTime,FLPar-method}
\title{Calculate recovery time to BMSY under Pella-Tomlinson dynamics}
\usage{
rTime(object, ...)

\S4method{rTime}{numeric}(object, r, p, bmsy = 1, F = 0, ...)

\S4method{rTime}{FLPar}(object, F = 0, ...)
}
\arguments{
\item{object}{Initial biomass (must be less than BMSY) or FLPar object with parameters}

\item{...}{Additional arguments}

\item{r}{Intrinsic rate of population increase (if object is numeric)}

\item{p}{Shape parameter of the Pella-Tomlinson model (if object is numeric)}

\item{bmsy}{Biomass at maximum sustainable yield (default=1)}

\item{F}{Fishing mortality rate during recovery (default=0)}
}
\value{
Recovery time in years
}
\description{
Numerically integrates the Pella-Tomlinson production function to estimate the time required 
for a population to recover from an initial biomass to BMSY, with optional fishing mortality during recovery.
}
\examples{
rTime(0.3, r=0.2, p=1.5)
rTime(0.3, r=0.2, p=1.5, F=0.05)

# With FLPar
library(FLCore)
params=FLPar(initial=0.3, r=0.2, p=1.5, bmsy=1)
rTime(params, F=0.05)

# Vectorized example with data frame (requires plyr and pellatParams function)
# dat <- mdply(expand.grid(fmsy=seq(0.05,0.5,length.out=21), shape=seq(0.1,0.6,length.out=21)),
#             function(fmsy, shape) pellatParams(fmsy=fmsy, bmsy=shape, virgin=1))
# dat <- transform(dat, recovery_time = rTime(0.3, r, p))
# head(dat)
# This now works without the error: 'the condition has length > 1' (vectorized support added)
}
