% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr-rs.R
\name{rvirgin}
\alias{rvirgin}
\alias{rvirgin,FLBRP,missing-method}
\alias{rvirgin,FLBRP,numeric-method}
\title{Recruitment at Virgin for a stock-recruitment model}
\usage{
rvirgin(object, ratio = 1, ...)

\S4method{rvirgin}{FLBRP,missing}(object, ratio = 1, ...)

\S4method{rvirgin}{FLBRP,numeric}(object, ratio = 1, ...)
}
\arguments{
\item{object}{An object (typically FLBRP)}

\item{ratio}{Optional numeric ratio (default 1.0)}

\item{...}{Additional arguments}
}
\value{
Recruitment at MSY (numeric or FLPar)
}
\description{
Recruitment at Virgin for a stock-recruitment model
}
