% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annualise.R, R/seasonalise.R
\docType{methods}
\name{annualise}
\alias{annualise}
\alias{seasonalise}
\title{annualise}
\usage{
annualise(x)

seasonalise(object, season = 1:4)
}
\arguments{
\item{object}{an \code{FLStock} object}

\item{seasons}{a numeric with seasons}
}
\value{
\code{FLStock} object

\code{FLStock} object
}
\description{
Takes an `FLStock` with seasons and collapses into years

While 'expand' adds seasons to an annually structured 'FLStock' it does not
  change the 'FLQuants' such as m, stock.wt, mat, stock.n, catch.n and harvest
  to take account of seasonal effects, such as growth and spawning. It also 
  does not account for changes in the population due to seasonal fishing. 
  Therefore 'seasonalise',  divides M into seasons, interpolates wts, and 
  projects a cohort across ages to estimate numbers and catch-at-age.
}
\examples{
\dontrun{
}
\dontrun{
data(ple4)
ple4seasonal=seasonalise(ple4)
plot(FLStocks("Annual"=ple4,"Seasonal"=ple4seasonal)
}
}
\seealso{
\code{\link{expand}}

\code{\link{expand}}
}
