% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pella_tomlinson.R
\name{show,PellaTomlinson-method}
\alias{show,PellaTomlinson-method}
\title{Show method for PellaTomlinson class}
\usage{
\S4method{show}{PellaTomlinson}(object)
}
\arguments{
\item{object}{PellaTomlinson object.}
}
\description{
Shows the summary of the PellaTomlinson object.
}
