% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-generic.R
\name{ssDiagnostics}
\alias{ssDiagnostics}
\title{Run Stock Synthesis diagnostics}
\usage{
ssDiagnostics(object, ...)
}
\arguments{
\item{object}{A Stock Synthesis object or path-like input, depending on method.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Method-specific diagnostics output.
}
\description{
Run Stock Synthesis diagnostics
}
