% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss-index-diagnostics.R
\name{ssIndexDiagnostics,data.frame-method}
\alias{ssIndexDiagnostics,data.frame-method}
\title{Build combined SS index diagnostics}
\usage{
\S4method{ssIndexDiagnostics}{data.frame}(object, ...)
}
\arguments{
\item{object}{A data.frame containing index observations.}

\item{...}{Additional arguments passed to `ssIndexScale` and `ssIndexRuns`.}
}
\value{
For `ssIndexDiagnostics`, a list with `scaled` and `runs`.
}
\description{
Build combined SS index diagnostics
}
