% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss-index-diagnostics.R
\name{ssIndexPrep,data.frame-method}
\alias{ssIndexPrep,data.frame-method}
\title{Prepare SS index data from a data.frame}
\usage{
\S4method{ssIndexPrep}{data.frame}(object, name = "name", year = "year", data = "data")
}
\arguments{
\item{object}{A data.frame containing index observations.}

\item{name}{Name of the index/series column.}

\item{year}{Name of the year column.}

\item{data}{Name of the observed index value column.}
}
\value{
For `ssIndexPrep`, a data.frame with `name`, `year`, and `data`.
}
\description{
Prepare SS index data from a data.frame
}
