% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-generic.R
\name{ssIndexPrep}
\alias{ssIndexPrep}
\title{Prepare SS index data for diagnostics}
\usage{
ssIndexPrep(object, ...)
}
\arguments{
\item{object}{Raw SS index object (e.g., list output from cpueSS) or a data.frame.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Method-specific standardized index input data.
}
\description{
Prepare SS index data for diagnostics
}
