% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-generic.R
\name{ssIndexRuns}
\alias{ssIndexRuns}
\title{Compute runs-test diagnostics for standardized SS index residuals}
\usage{
ssIndexRuns(object, ...)
}
\arguments{
\item{object}{A data.frame containing standardized index residuals.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Method-specific runs-test output.
}
\description{
Compute runs-test diagnostics for standardized SS index residuals
}
