% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refpt-processing.R
\name{standardizeRefptData}
\alias{standardizeRefptData}
\title{Standardize Reference Point Data}
\usage{
standardizeRefptData(refpt_data, remove_na = TRUE)
}
\arguments{
\item{refpt_data}{Raw reference point data}

\item{remove_na}{Logical, whether to remove NA values (default: TRUE)}
}
\value{
Cleaned and standardized data
}
\description{
Clean and standardize reference point data.
}
