% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pella_tomlinson.R
\name{summary,PellaTomlinson-method}
\alias{summary,PellaTomlinson-method}
\title{Summary method for PellaTomlinson class}
\usage{
\S4method{summary}{PellaTomlinson}(object, ...)
}
\arguments{
\item{object}{PellaTomlinson object.}

\item{...}{Additional arguments.}
}
\description{
Provides a summary of the PellaTomlinson object, including model parameters,
derived reference points, and stock data.
}
