% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pellatParams.R
\name{summary,pellatParams-method}
\alias{summary,pellatParams-method}
\title{Summary method for pellatParams}
\usage{
\S4method{summary}{pellatParams}(object, ...)
}
\arguments{
\item{object}{A pellatParams object}

\item{...}{Additional arguments (not used)}
}
\value{
A summary object
}
\description{
Summary method for pellatParams
}
\examples{
pt = pellatParams(c(msy = 500, fmsy = 0.3, virgin = 10000))
pt = estimateParams(pt)
summary(pt)
}
