% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{toLogits}
\alias{toLogits}
\title{Convert Steepness to Logit Scale}
\usage{
toLogits(h)
}
\arguments{
\item{h}{Numeric vector of steepness values (0.2 to 1.0)}
}
\value{
Numeric vector of logit-transformed steepness values
}
\description{
Converts steepness parameter to logit-transformed scale for numerical
stability in optimization routines.
}
\examples{
# Convert steepness values to logit scale for optimization
steepness_values <- c(0.3, 0.5, 0.8)
logit_values <- toLogits(steepness_values)
print(logit_values)

}
