% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{tryIt}
\alias{tryIt}
\title{Safe try wrapper}
\usage{
tryIt(x)
}
\arguments{
\item{x}{Expression to try}
}
\value{
Result of x if successful, NULL if error
}
\description{
Wraps a function call in try() and returns NULL on error instead of erroring
}
\keyword{internal}
