% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refs.R
\name{updateRefs}
\alias{updateRefs}
\title{Update Reference Points for FLBRP object}
\usage{
updateRefs(x)
}
\arguments{
\item{x}{An FLBRP object}
}
\value{
An FLBRP object with updated reference points
}
\description{
Update Reference Points for FLBRP object
}
