% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario-management.R
\name{validateBootstrapConfig}
\alias{validateBootstrapConfig}
\title{Validate Bootstrap Configuration}
\usage{
validateBootstrapConfig(bootstrap_config)
}
\arguments{
\item{bootstrap_config}{Bootstrap configuration}
}
\value{
Validation results
}
\description{
Validate bootstrap configuration parameters.
}
