% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-management.R
\name{validateFLBRPObject}
\alias{validateFLBRPObject}
\title{Validate FLBRP Object Structure}
\usage{
validateFLBRPObject(flbrp, required_components = c("refpts", "params"))
}
\arguments{
\item{flbrp}{FLBRP object to validate}

\item{required_components}{Character vector of required components}
}
\value{
Validation status
}
\description{
Validate FLBRP object structure and contents.
}
