% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel-utils.R
\name{validateInputData}
\alias{validateInputData}
\title{Validate Input Data for Analysis Functions}
\usage{
validateInputData(data, required_cols = NULL, data_type = "data")
}
\arguments{
\item{data}{Input data object}

\item{required_cols}{Character vector of required column names}

\item{data_type}{Character string describing the type of data}
}
\value{
Validation results
}
\description{
Validate input data for analysis functions.
}
