% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pellatUtils.R, R/pellatParams.R
\name{validatePellatParams}
\alias{validatePellatParams}
\title{Validate Pella-Tomlinson parameters}
\usage{
validatePellatParams(object)

validatePellatParams(object)
}
\arguments{
\item{object}{A pellatParams object}

\item{r}{Intrinsic growth rate}

\item{p}{Shape parameter}

\item{virgin}{Virgin biomass}

\item{bmsy}{Biomass at MSY}

\item{fmsy}{Fishing mortality at MSY}

\item{msy}{Maximum sustainable yield}
}
\value{
Logical indicating if parameters are valid

TRUE if valid, character vector of error messages if invalid
}
\description{
Validates a pellatParams object and returns detailed error messages if invalid.
}
\examples{
pt = pellatParams(c(msy = 500, fmsy = 0.3, virgin = 10000))
validatePellatParams(pt)
}
