% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario-management.R
\name{validatePriorConfig}
\alias{validatePriorConfig}
\title{Validate Prior Configuration}
\usage{
validatePriorConfig(prior_config)
}
\arguments{
\item{prior_config}{Prior configuration}
}
\value{
Validation results
}
\description{
Validate prior configuration parameters.
}
