% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-generic.R
\name{vBio}
\alias{vBio}
\title{Estimate vulnerable biomass from SS report components}
\usage{
vBio(object, ...)
}
\arguments{
\item{object}{A Stock Synthesis output object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Method-specific vulnerable biomass output.
}
\description{
Estimate vulnerable biomass from SS report components
}
