% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abi.R
\name{ABIMSY}
\alias{ABIMSY}
\title{Age-Based Indicator at MSY (Backwards-Compatible Wrapper)}
\usage{
ABIMSY(y, ref = "msy", p = 0.9)
}
\arguments{
\item{y}{An FLBRP object}

\item{ref}{Character string specifying the reference point (default = "msy")}

\item{p}{Numeric value between 0 and 1 (default = 0.9)}
}
\value{
An FLQuant object containing proportions at MSY
}
\description{
Backwards-compatible wrapper for \code{abiMsy()} function.
}
\seealso{
\code{\link{abiMsy}}
}
