% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abi.R
\name{ABIstock}
\alias{ABIstock}
\title{Age-Based Indicator for Stock (Backwards-Compatible Wrapper)}
\usage{
ABIstock(x, A)
}
\arguments{
\item{x}{An FLStock object}

\item{A}{An FLQuant object containing threshold age(s)}
}
\value{
An FLQuant object containing proportions
}
\description{
Backwards-compatible wrapper for \code{abi(x, A)} when A is an FLQuant.
}
\seealso{
\code{\link{abi}}
}
