% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebuild-pellaT.R
\name{brebuild}
\alias{brebuild}
\alias{brebuild,numeric-method}
\alias{brebuild,FLPar-method}
\title{Find initial biomass for a given recovery time to BMSY}
\usage{
brebuild(object, ...)

\S4method{brebuild}{numeric}(object, r, p, bmsy = 1, F = 0, ...)

\S4method{brebuild}{FLPar}(object, F = 0, ...)
}
\arguments{
\item{object}{Desired recovery time in years or FLPar object with parameters}

\item{...}{Additional arguments}

\item{r}{Intrinsic rate of population increase (if object is numeric)}

\item{p}{Shape parameter of the Pella-Tomlinson model (if object is numeric)}

\item{bmsy}{Biomass at maximum sustainable yield (default=1)}

\item{F}{Fishing mortality rate during recovery (default=0)}
}
\value{
Initial biomass
}
\description{
For a specified number of years, returns the biomass from which it takes exactly that time
to recover to BMSY under Pella-Tomlinson dynamics with optional fishing mortality.
}
\examples{
brebuild(3, r=0.2, p=1.5)
brebuild(3, r=0.2, p=1.5, F=0.05)

# With FLPar
params=FLPar(nyrs=3, r=0.2, p=1.5, bmsy=1)
brebuild(params, F=0.05)
}
