% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebiomass.R
\name{ebiomass}
\alias{ebiomass}
\alias{ebiomass,FLStock-method}
\title{Calculate Exploitable Biomass}
\usage{
ebiomass(object)

\S4method{ebiomass}{FLStock}(object)
}
\arguments{
\item{object}{An object of class FLQuant or data.frame (depends on the method).}

\item{pars}{A parameter object (class FLPar) containing necessary parameters.}

\item{lc}{A threshold value for lc.}

\item{lmax}{A threshold value for lmax.}

\item{...}{Additional arguments .}
}
\value{
An FLQuant object containing the exploitable biomass time series

estimates of Z.
}
\description{
Calculates the exploitable biomass from an FLStock object using selectivity-weighted
catch weights and stock numbers.

This is a generic S4 method that calculates something using the `haupt` function.
The specific implementation depends on the class of the input object.
}
\details{
Exploitable biomass is based on weighting catch weights by selectivity normalised 
by peak selectivity, then multiplying by stock numbers and summing across ages
}
\examples{
\dontrun{
data(ple4)
eb <- ebiomass(ple4)
}

}
