% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{FLRebuild}
\alias{FLRebuild}
\alias{FLRebuild-package}
\title{FLRebuild: Fisheries Stock Rebuilding Analysis}
\description{
A package for analyzing fisheries stock rebuilding trajectories 
and calculating rebuilding times. Provides methods for projecting stock 
rebuilding from different initial depletion levels using FLR 
(Fisheries Library for R) objects.
}
\examples{
# Load the package
library(FLRebuild)

# Example with biodyn object
# bd <- biodyn(FLPar(r=0.5, k=1000, p=1))
# rebuild_data <- rebuildTime(bd)

# Example with FLBRP object  
# eq <- lhEql(lhPar(FLPar(linf=250, s=0.9)))
# stk <- rebuild(eq)
}
\author{
\strong{Maintainer}: Author Name \email{author@example.com}

}
