% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{interp}
\alias{interp}
\title{Interpolate rebuilding times}
\usage{
interp(df)
}
\arguments{
\item{df}{Data frame with columns: initial, year, biomass}
}
\value{
Data frame with columns: initial, year
}
\description{
Finds the first year where biomass reaches the target (>= 1) for each initial depletion level
}
