% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/FLBRP-blim.R
\name{msyVirgin}
\alias{msyVirgin}
\alias{msyVirgin,FLBRP-method}
\title{Calculate MSY and Virgin State Metrics}
\usage{
msyVirgin(object, ...)

\S4method{msyVirgin}{FLBRP}(object, ...)
}
\arguments{
\item{object}{An object of class FLBRP.}

\item{...}{Additional arguments}
}
\value{
A named vector containing metrics for virgin and MSY states.
}
\description{
Generic for calculating key metrics (F, SSB, catch, ebiomass) at virgin and MSY states for an FLBRP object.
}
\seealso{
\code{\link{FLBRP}} \code{\link{refpts}} \code{\link{computeRefpts}}
}
