% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{msyVirgin}
\alias{msyVirgin}
\title{Calculate MSY and Virgin State Metrics}
\usage{
msyVirgin(object, ...)
}
\arguments{
\item{object}{An object of class FLBRP.}
}
\value{
A named vector containing metrics for virgin and MSY states.
}
\description{
Generic for calculating key metrics (F, SSB, catch, ebiomass) at virgin and MSY states for an FLBRP object.
}
\note{
No method is currently implemented for this generic.
}
