% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pella_tomlinson.R
\name{production}
\alias{production}
\title{Production Function}
\usage{
production(biomass, params, ...)

production(biomass, params, ...)
}
\arguments{
\item{biomass}{Numeric vector or FLQuant object.}

\item{params}{FLPar object.}

\item{...}{Additional arguments.}
}
\value{
Numeric vector of production values.

Numeric vector or FLQuant object.
}
\description{
Calculates the production (yield) given biomass and model parameters.

Generic function for calculating production.
}
\examples{
biomass = seq(0, 1, length.out = 101)
params = c(r = 0.5, p = 0.25, virgin = 1.0)
yield = production(biomass, params)

}
