% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/rebuild-methods.R,
%   R/FLBRP-rebuild.R
\name{rebuild}
\alias{rebuild}
\alias{rebuild,FLBRP-method}
\title{Rebuild a fish population}
\usage{
rebuild(
  object,
  targetF = NULL,
  targetSSB = NULL,
  nInitial = 100,
  growthRate = 0.25,
  minVal = 1e-06,
  maxVal = 1,
  burnin = 20,
  truncate = TRUE
)

\S4method{rebuild}{FLBRP}(
  object,
  targetF = computeRefpts(object)["msy", "harvest"] * 0,
  targetSSB = computeRefpts(object)["msy", "ssb"],
  nInitial = 100,
  growthRate = 0.25,
  minVal = 1e-06,
  maxVal = 1,
  burnin = 20,
  truncate = TRUE
)

\S4method{rebuild}{FLBRP}(
  object,
  targetF = computeRefpts(object)["msy", "harvest"] * 0,
  targetSSB = computeRefpts(object)["msy", "ssb"],
  nInitial = 100,
  growthRate = 0.25,
  minVal = 1e-06,
  maxVal = 1,
  burnin = 20,
  truncate = TRUE
)
}
\arguments{
\item{object}{An object representing the population}

\item{targetF}{Target fishing mortality during rebuilding}

\item{targetSSB}{Target spawning stock biomass}

\item{nInitial}{Number of initial SSB levels}

\item{growthRate}{Growth rate for depletion sequence}

\item{minVal}{Minimum depletion value}

\item{maxVal}{Maximum depletion value}

\item{burnin}{Number of years for burn-in period}

\item{truncate}{Whether to remove burn-in period}
}
\value{
An object with rebuilding trajectories
}
\description{
Projects rebuilding trajectories from different initial SSB levels
}
