% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/rebuild-methods.R,
%   R/biodyn-rebuild.R, R/FLBRP-rebuild.R
\name{rebuild}
\alias{rebuild}
\alias{rebuild,FLBRP-method}
\alias{rebuild,numeric-method}
\alias{rebuildTime,FLStock-method}
\title{Rebuild a fish population}
\usage{
rebuild(object, ...)

\S4method{rebuild}{FLBRP}(
  object,
  targetF = computeRefpts(object)["msy", "harvest"] * 0,
  targetSSB = computeRefpts(object)["msy", "ssb"],
  nInitial = 100,
  growthRate = 0.25,
  minVal = 1e-06,
  maxVal = 1,
  burnin = 20,
  truncate = TRUE
)

gt(x)

\S4method{rebuild}{numeric}(object, ...)

\S4method{rebuildTime}{FLStock}(object, nx = 101)
}
\arguments{
\item{object}{An FLStock object}

\item{targetF}{Target fishing mortality during rebuilding}

\item{targetSSB}{Target spawning stock biomass}

\item{nInitial}{Number of initial SSB levels}

\item{growthRate}{Growth rate for depletion sequence}

\item{minVal}{Minimum depletion value}

\item{maxVal}{Maximum depletion value}

\item{burnin}{Number of years for burn-in period}

\item{truncate}{Whether to remove burn-in period}

\item{x}{An object of class \code{Rebuild}}

\item{nx}{Number of interpolation points (default = 101)}

\item{r}{Numeric, intrinsic rate of increase}

\item{p}{Numeric, shape parameter}

\item{k}{Numeric, carrying capacity (default: 1e3)}

\item{b0}{Numeric, initial biomass ratio (default: 1)}

\item{nyrs}{Numeric, number of projection years (default: 50)}

\item{niters}{Numeric, number of iterations (default: 101)}
}
\value{
An object with rebuilding trajectories

A data frame containing:
  \item{shape}{Shape parameter at BMSY}
  \item{year}{Year when target is reached}
  \item{initial}{Initial biomass ratio}

Numeric value representing generation time

A data frame with columns year and initial
}
\description{
Projects rebuilding trajectories from different initial SSB levels

Performs rebuilding analysis for fish populations

Calculates generation time as 1/r
}
\examples{
\dontrun{
res <- rebuild(r=0.5, p=3, k=1000, b0=1)
}

}
