% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-generic.R, R/rebuild_methods.R
\name{rebuildTime}
\alias{rebuildTime}
\alias{rebuildTime,FLStock-method}
\alias{rebuildTime,biodyn-method}
\title{Calculate rebuilding time}
\usage{
rebuildTime(object, ...)

\S4method{rebuildTime}{FLStock}(object)

\S4method{rebuildTime}{biodyn}(
  object,
  target = refpts(object)["bmsy"],
  initial = 100,
  growthRate = 0.3,
  minVal = 1e-06,
  maxVal = 1,
  nx = 101
)
}
\arguments{
\item{object}{An object containing rebuilding trajectories (FLStock, biodyn, or FLPar)}

\item{...}{Additional arguments}
}
\value{
A data frame with columns year and initial (or T_recover and initial)
}
\description{
Calculate rebuilding time
}
\examples{
\dontrun{
# Example with FLStock object
library(FLCore)
data(ple4)
rebuild_times <- rebuildTime(ple4)

# Example with biodyn object
library(mpb)
params <- FLPar(r=0.5, k=1000, p=1)
bd <- biodyn(params)
rebuild_times <- rebuildTime(bd)
}

}
