% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{rebuildTime}
\alias{rebuildTime}
\title{Calculate rebuilding time}
\usage{
rebuildTime(object, nx = 101)
}
\arguments{
\item{object}{An object containing rebuilding trajectories}

\item{nx}{Number of interpolation points}
}
\value{
A data frame with columns year and initial
}
\description{
Calculate rebuilding time
}
\examples{
# Example with FLStock object
# rebuild_times <- rebuildTime(stk)
} 