% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pellatRebuild.R
\name{rTime}
\alias{rTime}
\alias{rTime,numeric-method}
\alias{rTime,FLPar-method}
\alias{rTime,FLBRP-method}
\title{Calculate recovery time to BMSY under Pella-Tomlinson dynamics}
\usage{
rTime(object, ...)

\S4method{rTime}{numeric}(object, r, p, bmsy = 1, F = 0, ...)

\S4method{rTime}{FLPar}(object, F = 0, ...)

\S4method{rTime}{FLBRP}(object, initial, F = 0, biomass = "ssb", ...)
}
\arguments{
\item{object}{Initial biomass (must be less than BMSY) or FLPar object with parameters}

\item{...}{Additional arguments}

\item{r}{Intrinsic rate of population increase (if object is numeric)}

\item{p}{Shape parameter of the Pella-Tomlinson model (if object is numeric)}

\item{bmsy}{Biomass at maximum sustainable yield (default=1)}

\item{F}{Fishing mortality rate during recovery (default=0)}
}
\value{
Recovery time in years
}
\description{
Numerically integrates the Pella-Tomlinson production function to estimate the time required 
for a population to recover from an initial biomass to BMSY, with optional fishing mortality during recovery.
}
\examples{
rTime(0.3, r=0.2, p=1.5)
rTime(0.3, r=0.2, p=1.5, F=0.05)

# With FLPar
library(FLCore)
params=FLPar(initial=0.3, r=0.2, p=1.5, bmsy=1)
rTime(params, F=0.05)

# With FLBRP object
data(ple4brp)
rTime(ple4brp, initial=0.3, F=0.05)
}
